==> analysis/analyses/experiment_n1008/3b_analyze_FLH.Rmd <==
---
title: "Action Asymmetry Experiment (n = 1008): Field by Level by Handedness Analyses"
pagetitle: "aah | exp | analysis (field by level by handedness)"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

<style type="text/css">
  body{
  font-family: Avenir;
  font-size: 12pt;
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.align = "center", fig.width = 9,
                      fig.height = 6, results = "asis")
options(knitr.kable.NA = "")

cli.progress_show_after <- 0

## Do not use scientific notation until 9 decimal places.
options(scipen = 9, digits = 9)

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)
```

```{r lib}
library(here)
library(tidyverse)
library(cli) # For printing error messages

library(lme4)
library(emmeans)
library(broom)
library(gt)

library(ggpubr) ## functions for plotting SD, SEM, CI.
library(ggbeeswarm)
library(patchwork)

library(knitr) # For include_graphics

source(here::here("lib", "load_process", "load_process.R"))
source(here::here("lib", "util.R"))
source(here::here("lib", "demographics.R"))
source(here::here("lib", "plots.R"))

## Colors for plots
plot_blue <- c("#337aB7")
plot_colors <- c("#4477AA",
                 "#EE6677",
                 "#CCBB44",
                 "#66CCEE",
                 "#AA3377",
                 "#228833",
                 "#BBBBBB")
```

```{r config}
proc_dir <- here::here("data", "proc_exp_n1008")
fig_dir <- here::here("figures", "exp_n1008_b")
manual_cache_dir <- here::here("manual_cache", "exp_n1008_b")

use_cached_model_rt <- T
use_cached_model_acc <- T
use_cached_model_rt_ehi <- T
use_cached_model_acc_ehi <- T

use_cached_figs <- T

if (use_cached_figs == TRUE) {
  use_cached_demofigs <- TRUE
  use_cached_rtfigs <- TRUE
  use_cached_accfigs <- TRUE
} else if (use_cached_figs == FALSE) {
  use_cached_demofigs <- FALSE
  use_cached_rtfigs <- FALSE
  use_cached_accfigs <- FALSE
}
```

# {.tabset}
```{r load_data}
## Load "the data" with all subjects & trials.
aah_long <- load_aah_long(proc_dir)

## Load summary data table (for quick demographic analyses)
## with all subjects.
aah_summary_all <- load_aah_summary(proc_dir)
```

```{r prepare_data}
#### PREPARE SUMMARY DATA FOR DEMOGRAPHICS ANALYSES
aah_summary <- filter_aah_summary_for_analysis(aah_summary_all)

#### PREPARE TRIAL-LEVEL DATA FOR RT, ACCURACY ANALYSES
## Prepare "the data" (aah_long) for all analyses:
## Filter out practice trials, absent trials, and excluded subjects
## This data (all present trials, correct or incorrect) will be used for
## accuracy analyses
aah <- filter_aah_long_for_analysis(aah_long)

## For RT analyses, prepare dataset with only correct, present trials.
## In our RT model, we only care about correct responses to present trials.
aah_correct <- aah |> filter(correct == T)
## Code handedness

## Relevel field and level with RVF first (unintuitive for plotting),
## so that emmeans will show a positive number for LVF global bias.
aah_for_rt_model <- aah_correct |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("RVF", "LVF")))

## For accuracy analyses, prepare dataset will all present trials.
## Relevel field and level,
## so that emmeans will show a positive number for LVF global bias.
aah_for_acc_model <- aah |>
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("RVF", "LVF")))

#### PREPARE SUBJECT_LEVEL DATA FOR RT, ACCURACY ANALYSES
## Prepare subject-level data (RT)
rt_subject <- aah_correct |> group_by(subject, field, level, handedness) |>
  summarize(rt = median(rt))

## Prepare subject-level LVF Global bias summary (RT)
rt_1 <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  mutate(all_one_group = "all_one_group")

## Prepare subject-level LVF Global bias summary (Accuracy)
## (For accuracy, include incorrect trials!)
acc_subject <- aah |> group_by(subject, field, level, handedness) |>
  summarize(
    total_responses = n(),
    n_present_resp = sum(correct),
    n_absent_resp = total_responses - n_present_resp,
    n_correct = sum(correct),
    acc = 100 * (n_correct / total_responses)
  ) |>
  select(subject, handedness, field, level, acc) |>
  mutate(level = recode(level, global = "Global", local = "Local"))

acc_1 <- acc_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = acc) |>
  mutate(LVF_Global_Bias = (LVF_Global - LVF_Local) - (RVF_Global - RVF_Local)) |>
  mutate(all_one_group = "all_one_group")
```

## Demographics {.tabset .tabset-pills}

### Everyone

Demographics for all included participants.
```{r demo_summary}
rt_demo <- demo_summary_table(aah_summary)
rt_demo |> pretty_table() |> tab_header(title = "Demographics", subtitle = "Summary")
```
<br>
```{r demo_plot_config}
plot_color = plot_blue
```

```{r demo_plot_age}
fig_path <- here(fig_dir, "demo_age.png")

if (use_cached_demofigs == FALSE) {
g <- ggplot(aah_summary, aes(x = age)) +
  geom_histogram(color = NA, fill = plot_color, alpha = .8, binwidth = 1) +
  labs(x = "Age (years)")

g <- g |> gg_style_demo() + theme(aspect.ratio = 1/2,
                            axis.title.y = element_blank())

ggsave(fig_path, g, "png", height = 2, width = 4)
}

include_graphics(fig_path)
```

```{r demo_plot_edu}
fig_path <- here(fig_dir, "demo_edu.png")

if (use_cached_demofigs == FALSE) {
g <- ggplot(aah_summary, aes(x = education)) +
  geom_histogram(color = NA, fill = plot_color, alpha = .8, binwidth = 1) +
  labs(x = "Education (years)")

g <- g |> gg_style_demo() + theme(aspect.ratio = 1/2,
                            axis.title.y = element_blank(),
                            axis.text.y = element_blank())

ggsave(fig_path, g, "png", height = 2, width = 4)
}

include_graphics(fig_path)
```
<br>
```{r demo_race}
demo_race_table(aah_summary) |> pretty_table()
```
<br>
```{r demo_ethnicity}
demo_ethnicity_table(aah_summary) |> pretty_table()
```
<br>

### By handedness group {.active}

Demographics for included participants, by handedness group (EHI bins). 
```{r}
hand_demo_summary_table(aah_summary) |> pretty_table() |>
  tab_footnote("Left: (EHI <= -40)  |  Mixed: (-40 < EHI < 40)  |  Right: (EHI >= 40)")
```
<br>
```{r demo_plot_ehi_binned}
fig_path <- here(fig_dir, "demo_ehi_binned.png")

if (use_cached_demofigs == FALSE) {
g <- ggplot(aah_summary, aes(x = ehi)) +
  geom_vline(xintercept = -36, color = "gray50") +
  geom_vline(xintercept = 36, color = "gray50") +
  geom_histogram(color = NA, fill = plot_color, alpha = .8, binwidth = 8, boundary = 100) +
  labs(x = "EHI")+
  xlim(c(-104, 104))

g <- g |> gg_style_demo() + theme(aspect.ratio = 1/2,
                            axis.title.y = element_blank())

ggsave(fig_path, g, "png", height = 2, width = 4)
}

include_graphics(fig_path)
```


## Field x Level x Handedness (binned) {.tabset .tabset-pills}

Do we find an interaction of field x level x handedness, when handedness is binned as left (EHI <= -40) or right (EHI > +40)?
<br>
<br>

 ***Summary.*** For reaction time, we find the critical interaction in the predicted direction (11.67ms, 95% CI [0.65, 22.69], p = .019, one-sided). Left handers show 15.64ms LVF global bias (95% CI [7.61, 23.67]), and right handers 27.31ms (95% CI [19.80, 34.81]). Mixed handers (not included in the categorical interaction analysis) show a LVF global bias of 21.66ms (95% CI [9.09, 34.23]).
 
For accuracy, we find no significant interaction of field by level by handedness (OR = 0.90, 95% CI [0.70, 1.16], p = .21, one-sided; where OR < 1 means greater LVF global bias for left handers). Point estimates of LVF global bias hardly differ between left handers (OR = 1.95, 95% CI [1.62, 2.35]) and right handers (OR = 1.77, 95% CI [1.49, 2.10]). For mixed handers (not included in the categorical interaction analysis), the point estimate is 1.10 (95% CI [0.82, 1.47])

<br>
<br>

### Reaction time {.tabset}

#### Plots

Error bars show 95% CI.
```{r plot_rt_config}
plot_color <- plot_colors[[5]] # Color for interaction plots with only one color.
```

```{r plot_rt_4a_all_trials}
## Plot data from every trial, with overall mean, median and bounds.
fig_path_rt_4a <- here(fig_dir, "rt_4a.png")

if (use_cached_rtfigs == FALSE) {
rt_descriptive <- aah_correct |>
  group_by(field, level, handedness) |>
  summarize(
    median = median(rt),
    mean = mean(rt),
    SE = sd(rt) / sqrt(length((rt)))
  )

## Facet by: Visual field, handedness
g <- ggplot(aah_correct, aes(
  x = level,
  y = rt,
  fill = level,
  color = level,
)) +
  geom_quasirandom(alpha = .05, show.legend = F) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .3,
    # size = point_size / 5,
    show.legend = F
  ) +
  geom_point(data = rt_descriptive, aes(y = mean), color = "black", shape = 21,
             show.legend = F) +
  scale_y_continuous(minor_breaks = seq(0 , 2000, 100),
                     breaks = seq(0, 2000, 200)) +
  facet_grid(vars(handedness), vars(field)) +
  labs(title = "All trials (RT)", x = "Level", y = "Reaction time (ms)")

g <- g |> gg_style_means() |> gg_color()
ggsave(fig_path_rt_4a, g, "png", height = 8, width = 4)
g_rt_4a <- g
}

include_graphics(fig_path_rt_4a)
```
<br>

---

<br>
```{r plot_rt_4}
fig_path_rt_4 <- here(fig_dir, "rt_4.png")

if (use_cached_rtfigs == FALSE) {
rt_subject <- aah_correct |>
  group_by(subject, field, level, handedness) |> 
  summarize(rt = median(rt))

rt_descriptive <- rt_subject |>
  group_by(field, level, handedness) |>
  summarize(
    median = median(rt),
    mean = mean(rt),
    SE = sd(rt) / sqrt(length((rt)))
  )

g <- ggplot(rt_subject, aes(
  x = level,
  y = rt,
  fill = level,
  color = level
)) +
  geom_quasirandom(alpha = .1, show.legend = F) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    # size = point_size / 5,
    show.legend = F
  ) +
  geom_point(
    data = rt_descriptive,
    aes(y = mean),
    color = "black",
    shape = 21,
    show.legend = F
  ) +
  scale_y_continuous(minor_breaks = seq(0 , 1500, 100),
                     breaks = seq(0, 1500, 200)) +
  facet_grid(vars(handedness), vars(field)) +
  labs(title = "Per-subject medians (RT)", x = "Level", y = "Reaction time (ms)")

g <- g |>  gg_style_means() |> gg_color()
ggsave(fig_path_rt_4, g, "png", height = 8, width = 4)
g_rt_4 <- g
}

include_graphics(fig_path_rt_4)
```
```{r plot_rt_4a4}
# fig_path_rt_4a4 <- here(fig_dir, "rt_combo_4a-4.png")
# 
# if (use_cached_rtfigs == FALSE) {
# g_rt_4a4 <- g_rt_4a + g_rt_4
# g_rt_4a4
# ggsave(fig_path_rt_4a4, g_rt_4a4, "png", height = 4, width = 8)
# }
# 
# include_graphics(fig_path_rt_4a4)
```

---

```{r plot_rt_2}
fig_path_rt_2 <- here(fig_dir, "rt_2.png")

if (use_cached_rtfigs == FALSE) {
## Make a table showing:
## For each subject and field, the difference in median rt for:
## Global - Local
rt_2 <- rt_subject |> 
  pivot_wider(names_from = c(level),
              values_from = rt) |> 
  mutate(Global_Bias = Local - Global)

rt_descriptive <- rt_2 |>
  group_by(field, handedness) |>
  summarize(
    median = median(Global_Bias),
    mean = mean(Global_Bias),
    SE = sd(Global_Bias) / sqrt(length((Global_Bias)))
  )

g <- ggplot(rt_2, aes(x = field,
                      y = Global_Bias, fill = handedness, color = handedness)) +
  geom_quasirandom(
    alpha = .1, show.legend = F,
    # color = plot_color
    ) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    # fill = plot_color,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    show.legend = F
  ) +
  geom_point(
    data = rt_descriptive,
    aes(y = mean),
    # fill = plot_color,
    color = "gray5",
    shape = 21,
    show.legend = F
  ) +
  facet_wrap(~handedness)+
  scale_y_continuous(minor_breaks = seq(-500 , 500, 50),
                     breaks = seq(-500, 500, 100)) +
  scale_fill_manual(values = plot_colors[c(1, 5,  2)]) +
  scale_color_manual(values = plot_colors[c(1, 5,  2)]) +
  labs(title = "Global bias (RT)", x = "Level", y = "Local - Global RT (ms)")

g <- g |> gg_style_means()
ggsave(fig_path_rt_2, g, "png", height = 4, width = 4)
g_rt_2 <- g
}

include_graphics(fig_path_rt_2)
```

---

```{r plot_rt_1}
fig_path_rt_1 <- here(fig_dir, "rt_1.png")

if (use_cached_rtfigs == FALSE) {
  ## Make a table showing:
  ## For each subject and field, the difference in median rt for:
  ## Global - Local
  rt_1 <- rt_subject |>
    pivot_wider(names_from = c(field, level),
                values_from = rt) |>
    mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
    mutate(all_one_group = "all_one_group")
    
rt_descriptive <- rt_1 |>
  group_by(all_one_group, handedness) |> 
  summarize(
    median = median(LVF_Global_Bias),
    mean = mean(LVF_Global_Bias),
    SE = sd(LVF_Global_Bias) / sqrt(length((LVF_Global_Bias)))
  )

g <- ggplot(rt_1, aes(
  x = all_one_group,
  y = LVF_Global_Bias,
  fill = handedness,
  color = handedness
)) +
  geom_quasirandom(
    alpha = .1, show.legend = F,
    # color = plot_color
    ) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    color = "gray20",
    # fill = plot_color,
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    # size = point_size / 5,
    show.legend = F
  ) +
  geom_point(
    data = rt_descriptive,
    aes(y = mean),
    # fill = plot_color,
    color = "black",
    shape = 21,
    show.legend = F
  ) +
  facet_wrap(~handedness) +
  scale_y_continuous(minor_breaks = seq(-500 , 500, 50),
                     breaks = seq(-500, 500, 100)) +
  scale_fill_manual(values = plot_colors[c(1, 5,  2)]) +
  scale_color_manual(values = plot_colors[c(1, 5,  2)]) +
  theme(axis.text.x = element_blank()) +
  labs(title = "LVF>RVF Global Bias (RT)", x = "Level", y = "RVF - LVF, Local - Global RT (ms)")

g <- g |> gg_style_means() +
  theme(axis.text.x = element_blank(), aspect.ratio = 4/1)
ggsave(fig_path_rt_1, g, "png", height = 4, width = 4)
g_rt_1 <- g
}

include_graphics(fig_path_rt_1)
```
```{r plot_rt_21}
# fig_path_rt_21 <- here(fig_dir, "rt_combo_2-1.png")
# 
# if (use_cached_rtfigs == FALSE) {
# g_rt_1_padded <- g_rt_1 + theme(plot.margin = unit(c(10, 50, 5.5, 40), "pt"))
# g_rt_21 <- g_rt_2 + g_rt_1_padded
# ggsave(fig_path_rt_21, g_rt_21, "png", height = 4, width = 8)
# }
# 
# include_graphics(fig_path_rt_21)
```

#### Statistics

##### Simple mixed regression model
Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness + (1 | subject) )`
<br>
<br>

```{r rt_model_2bins}
## Make a model with two handedness bins: Right and Left
aah_for_rt_model_2bins <- aah_for_rt_model |> filter(handedness %in% c("Right", "Left"))
rt_model_2bins <- lmer(rt ~ field*level*handedness + (1 | subject), data = aah_for_rt_model_2bins)

if (use_cached_model_rt == FALSE) {
  ## Create emmeans model object, and manually cache it.
  rt_emm_2bins <- emmeans(rt_model_2bins, ~ field * level * handedness)
  ## Todo: run without z-approximation:
  # rt_emm <- emmeans(rt_model, ~ field * level * handedness, pbkrtest.limit = 59851)
  
  ## Manually cache model
  saveRDS(rt_emm_2bins, here(manual_cache_dir, "rt_emm_2bins.rds"))
  
} else if (use_cached_model_rt == TRUE) {
  ## Load cached model
  rt_emm_2bins <- readRDS(here(manual_cache_dir, "rt_emm_2bins.rds"))
}
```
```{r}
## Make a model with three handedness bins: Right and Left
rt_model_3bins <- lmer(rt ~ field*level*handedness + (1 | subject), data = aah_for_rt_model)

if (use_cached_model_rt == FALSE) {
  ## Create emmeans model object, and manually cache it.
  rt_emm_3bins <- emmeans(rt_model_3bins, ~ field * level * handedness)
  ## Todo: run without z-approximation:
  # rt_emm <- emmeans(rt_model, ~ field * level * handedness, pbkrtest.limit = 59851)
  
  ## Manually cache model
  saveRDS(rt_emm_3bins, here(manual_cache_dir, "rt_emm_3bins.rds"))
  
} else if (use_cached_model_rt == TRUE) {
  ## Load cached model
  rt_emm_3bins <- readRDS(here(manual_cache_dir, "rt_emm_3bins.rds"))
}
```


<br>
```{r rt_interaction_anova}
## Use anova() on competing models to test 2-way interaction.
interaction_stats <-
  function(model_with_interaction,
           model_with_no_interaction) {
    return(anova(model_with_interaction, model_with_no_interaction))
  }

rt_model_no_interaction <- update(rt_model_2bins, . ~ . - field:level:handedness)
interaction_anova <- interaction_stats(rt_model_2bins, rt_model_no_interaction)
interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level by handedness interaction (RT)", 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
```
<br>
```{r rt_interaction_emm}
## Use emmeans() to test 3-way interaction.
rt_interaction_emm <- rt_emm_2bins |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "Field by level by handedness interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means LVF global bias is stronger in right handers (as predicted by AAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r rt_handedness_bias}
## Estimate the effect of field by level for each handedness group
rt_emm2 <- emmeans(rt_model_3bins, ~field*level | handedness) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")

rt_emm2 |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "LVF Global bias by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r rt_interacion_aov}
## Use aov to test 2-way interaction with a traditional F-test.
rt_aov <- aov(rt ~ field*level*handedness, data = aah_for_rt_model_2bins)
rt_aov_summary <- summary(rt_aov)
rt_aov_summary |> (\(.) .[[1]])() |> tidy() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "Field by level interaction (RT)",
             subtitle = "Old-school Omnibus F-test")
```
<br>

---

<br>
```{r, echo = T, results = "markup"}
summary(rt_model_2bins)
```
<br>


### Accuracy {.tabset .active}

#### Plots

Error bars show 95% CI.
```{r plot_acc_config}
plot_color <- plot_colors[[5]] # Color for interaction plots with only one color.
```
<br>
```{r plot_acc_4}
fig_path_acc_4 <- here(fig_dir, "acc_4.png")

if (use_cached_accfigs == FALSE) {
  acc_subject

acc_descriptive <- acc_subject |>
  group_by(field, level, handedness) |>
  summarize(
    median = median(acc),
    mean = mean(acc),
    SE = sd(acc) / sqrt(length((acc)))
  )

g <- ggplot(acc_subject, aes(
  x = level,
  y = acc,
  fill = level,
  color = level
)) +
  geom_quasirandom(alpha = .1, show.legend = F) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    # size = point_size / 5,
    show.legend = F
  ) +
  geom_point(
    data = acc_descriptive,
    aes(y = mean),
    color = "black",
    shape = 21,
    show.legend = F
  ) +
  scale_y_continuous(minor_breaks = seq(-100 , 100, 10),
                     breaks = seq(-100, 100, 20)) +
  facet_grid(vars(handedness), vars(field)) +
  labs(title = "Per-subject accuracy", x = "Level", y = "Accuracy (% correct)")

g <- g |>  gg_style_means() |> gg_color()
ggsave(fig_path_acc_4, g, "png", height = 8, width = 4)
g_acc_4 <- g
}

include_graphics(fig_path_acc_4)
```

---

```{r plot_acc_2}
fig_path_acc_2 <- here(fig_dir, "acc_2.png")

if (use_cached_accfigs == FALSE) {
## Make a table showing:
## For each subject and field, the difference in median acc for:
## Global - Local
acc_2 <- acc_subject |> 
  pivot_wider(names_from = c(level),
              values_from = acc) |> 
  mutate(Global_Bias = Global - Local)

acc_descriptive <- acc_2 |>
  group_by(field, handedness) |>
  summarize(
    median = median(Global_Bias),
    mean = mean(Global_Bias),
    SE = sd(Global_Bias) / sqrt(length((Global_Bias)))
  )

g <- ggplot(acc_2, aes(x = field,
                      y = Global_Bias, fill = handedness, color = handedness)) +
  geom_quasirandom(
    alpha = .1, show.legend = F,
    # color = plot_color
    ) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    # fill = plot_color,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    show.legend = F
  ) +
  geom_point(
    data = acc_descriptive,
    aes(y = mean),
    # fill = plot_color,
    color = "gray5",
    shape = 21,
    show.legend = F
  ) +
  facet_wrap(~handedness)+
  scale_y_continuous(minor_breaks = seq(-100 , 100, 10),
                     breaks = seq(-100, 100, 20)) +
  scale_fill_manual(values = plot_colors[c(1, 5,  2)]) +
  scale_color_manual(values = plot_colors[c(1, 5,  2)]) +
  labs(title = "Global bias (Accuracy)", x = "Level", y = "Global - Local Accuracy (% correct)")

g <- g |> gg_style_means()
ggsave(fig_path_acc_2, g, "png", height = 4, width = 4)
g_acc_2 <- g
}

include_graphics(fig_path_acc_2)
```

---

```{r plot_acc_1}
fig_path_acc_1 <- here(fig_dir, "acc_1.png")

if (use_cached_accfigs == FALSE) {
  ## Make a table showing:
  ## For each subject and field, the difference in median acc for:
  ## Global - Local
  acc_1 <- acc_subject |>
    pivot_wider(names_from = c(field, level),
                values_from = acc) |>
    mutate(LVF_Global_Bias = (LVF_Global - LVF_Local) - (RVF_Global - RVF_Local)) |>
    mutate(all_one_group = "all_one_group")
    
acc_descriptive <- acc_1 |>
  group_by(all_one_group, handedness) |> 
  summarize(
    median = median(LVF_Global_Bias),
    mean = mean(LVF_Global_Bias),
    SE = sd(LVF_Global_Bias) / sqrt(length((LVF_Global_Bias)))
  )

g <- ggplot(acc_1, aes(
  x = all_one_group,
  y = LVF_Global_Bias,
  fill = handedness,
  color = handedness
)) +
  geom_quasirandom(
    alpha = .1, show.legend = F,
    # color = plot_color
    ) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    color = "gray20",
    # fill = plot_color,
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    # size = point_size / 5,
    show.legend = F
  ) +
  geom_point(
    data = acc_descriptive,
    aes(y = mean),
    # fill = plot_color,
    color = "black",
    shape = 21,
    show.legend = F
  ) +
  facet_wrap(~handedness) +
  # scale_y_continuous(minor_breaks = seq(-500 , 500, 50),
  #                    breaks = seq(-500, 500, 100)) +
  scale_fill_manual(values = plot_colors[c(1, 5,  2)]) +
  scale_color_manual(values = plot_colors[c(1, 5,  2)]) +
  theme(axis.text.x = element_blank()) +
  labs(title = "LVF>RVF Global Bias (Accuracy)", x = "Level", y = "LVF - RVF, Global - Local Accuracy (% correct)")

g <- g |> gg_style_means() +
  theme(axis.text.x = element_blank(), aspect.ratio = 4/1)
ggsave(fig_path_acc_1, g, "png", height = 4, width = 4)
g_acc_1 <- g
}

include_graphics(fig_path_acc_1)
```

#### Statistics
##### Simple mixed regression model

Accuracy is modeled as a binomial effect of field, level, and handedness, using binary correct/incorrect data from every target-present trial:
<br>
<br>
`glmer( correct ~ field*level*handedness + (1 | subject), family = "binomial" )`
<br>
<br>
```{r acc_model_2bins}
## Make a binomial logistic model using data from every trial.


if (use_cached_model_acc == FALSE) {
  ## Make a model with two handedness bins: Right and Left
  aah_for_acc_model_2bins <-
    aah_for_acc_model |> filter(handedness %in% c("Right", "Left"))
  
  acc_model_2bins <-
    glmer(correct ~ field * level * handedness + (1 | subject),
          data = aah_for_acc_model_2bins,
          family = "binomial")
  
  ## Warning: the model did not converge
  ## max|grad .006 (tol .002)
  ## https://rstudio-pubs-static.s3.amazonaws.com/33653_57fc7b8e5d484c909b615d8633c01d51.html
  
  
  ## I re-ran the model with the bobyqa optimizer on Jan 7 2023,
  ## and saw virtually the same estimates.
  # acc_model_2bins <-
  #   glmer(correct ~ field * level * handedness + (1 | subject),
  #         data = aah_for_acc_model_2bins,
  #         family = "binomial",
  #         control = glmerControl(optimizer = c("bobyqa"))
  #         )
  
  
  ## Create emmeans model object, and manually cache it.
  acc_emm_2bins <-
    emmeans(acc_model_2bins, ~ field * level & handedness)
  
  ## Manually cache model
  saveRDS(acc_model_2bins,
          here(manual_cache_dir,
               "acc_model_2bins.rds"))
  saveRDS(acc_emm_2bins, here(manual_cache_dir, "acc_emm_2bins.rds"))
  
} else if (use_cached_model_acc == TRUE) {
  ## Load cached model
  acc_model_2bins <- readRDS(here(manual_cache_dir, "acc_model_2bins.rds"))
  acc_emm_2bins <- readRDS(here(manual_cache_dir, "acc_emm_2bins.rds"))
}
```


```{r}
if (use_cached_model_acc == FALSE) {
  ## Make a model with three handedness bins: Right, Mixed, and Left
  acc_model_3bins <-
    glmer(correct ~ field * level * handedness + (1 | subject),
          data = aah_for_acc_model,
          family = "binomial")
  
  ## Warning message:
  ##In checkConv(attr(opt, "derivs"), opt$par, ctrl = control$checkConv,  :
  ##  Model failed to converge with max|grad| = 0.130807 (tol = 0.002, component 1)
  
  ## Create emmeans model object, and manually cache it.
  acc_emm_3bins <-
    emmeans(acc_model_3bins, ~ field * level * handedness)
  
  ## Manually cache model
  saveRDS(acc_model_3bins,
          here(manual_cache_dir, "acc_model_3bins.rds"))
  saveRDS(acc_emm_3bins, here(manual_cache_dir, "acc_emm_3bins.rds"))
  
} else if (use_cached_model_acc == TRUE) {
  ## Load cached model
  acc_model_3bins <-
    readRDS(here(manual_cache_dir, "acc_model_3bins.rds"))
  acc_emm_3bins <-  readRDS(here(manual_cache_dir, "acc_emm_3bins.rds"))
}
```

<br>
```{r acc_interaction_anova}
## Use anova() on competing models to test 2-way interaction.
if (use_cached_model_acc == FALSE) {
  ## Create and manually cache ANOVA
  acc_model_no_interaction <-
    update(acc_model_2bins, . ~ . - field:level:handedness)
  
  interaction_anova <-
    interaction_stats(acc_model_2bins, acc_model_no_interaction)
  
  ## Manually cache model
  saveRDS(interaction_anova, here(manual_cache_dir,
                                "acc_anova_2bins.rds"))
  
} else if (use_cached_model_acc == TRUE) {
  ## Load cached ANOVA
  interaction_anova <- readRDS(here(manual_cache_dir, "acc_anova_2bins.rds"))
}

interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level by handedness interaction (Accuracy)", 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
```
<br>
```{r acc_interaction_emm}
## Use emmeans() to test 3-way interaction.
acc_interaction_emm <- acc_emm_2bins |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T, type = "response")

acc_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "Field by level by handedness interaction (Accuracy)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote =  "Backtransformed to odds ratio from log odds ratio (tests are performed on log odds ratio scale). A ratio > 1 means global bias is stronger in the LVF for right handers (predicted by AAH)",
               locations = cells_column_labels(columns = odds.ratio)) |>
  tab_footnote(footnote = "'Inf' df is expected when emmeans does logistic regression. See emmeans FAQ: https://cran.r-project.org/web/packages/emmeans/vignettes/FAQs.html#asymp.",
               locations = cells_column_labels(columns = df)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value))
```
<br>
```{r acc_handedness_bias}
## Estimate the effect of field by level for each handedness group
acc_emm2 <- emmeans(acc_model_3bins, ~field*level | handedness) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none", type = "response")

acc_emm2 |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "LVF Global bias by handedness bin (Accuracy)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote =  "Backtransformed to odds ratio from log odds ratio (tests are performed on log odds ratio scale). A ratio > 1 means global bias is stronger in the LVF, as predicted for right handers. Mixed handers' global bias is shown here, but their data was not included in the binomial model.",
               locations = cells_column_labels(columns = odds.ratio)) |>
  tab_footnote(footnote = "'Inf' df is expected when emmeans does logistic regression. See emmeans FAQ: https://cran.r-project.org/web/packages/emmeans/vignettes/FAQs.html#asymp.",
               locations = cells_column_labels(columns = df)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value))
```
<br>

---

<br>
```{r, echo = T, results = "markup"}
summary(acc_model_2bins)
```
<br>
Using the bobyqa optimizer alone gives virtually the same results.
```{r}
acc_model_2bins_bobyqa <- readRDS(here(manual_cache_dir, "models_bobyqa", 
                                     "acc_model_2bins.rds"))
```
```{r, echo = T, results = "markup"}
summary(acc_model_2bins_bobyqa)
```
<br>

## Field x Level x Handedness (continuous) {.tabset .tabset-pills}

Do we find an interaction of field x level x handedness (continuous EHI score)?
<br>
<br>

***Summary.*** For reaction time, we find the critical interaction in the predicted direction (.067ms per EHI unit, 95% CI [0.003, 0.13], p = .020, one-sided). Estimated global bias is 13.32ms lower for strong left handers (EHI -100: 14.82ms, 95% CI [6.48, 23.17]) than for strong right handers (EHI +100: 28.14ms, 95% CI [20.31, 35.98]).

For accuracy, we find no significant interaction of field by level by EHI (Beta = 0.0002 logodds per EHI unit, 95% CI [-0.001, 0.002], p = .40, one-sided). Estimated LVF global bias hardly differs for strong left handers (EHI -100: OR = 1.67, 95% CI [1.37, 2.03]) and strong right handers (EHI +100: OR = 1.73, 95% CI [1.45, 2.07]).

<br>
<br>

### Reaction time
```{r}
rt_ehi_1 <- rt_1 |>
  left_join(aah_summary |> select(subject, ehi)) |>
  select(subject, ehi, LVF_Global_Bias)
```


#### Plots
```{r}
fig_path_rt_ehi_cor <- here(fig_dir, "rt_ehi_cor.png")
fig_path_rt_ehi_cor_wide <- here(fig_dir, "rt_ehi_cor_wide.png")
g <- rt_ehi_1 |> ggplot(aes(x = ehi, y = LVF_Global_Bias)) +
  geom_beeswarm(alpha = .2) +
  geom_smooth(method = "lm", color = "black") +
  labs(x = "EHI", y = "RVF - LVF, Local - Global RT (ms)",
       title = "EHI and each subject's mean LVF global bias")

g <- g |> gg_style()
ggsave(fig_path_rt_ehi_cor, g, "png", height = 4, width = 4)
ggsave(fig_path_rt_ehi_cor_wide, g, "png", height = 4, width = 8)
```
```{r}
include_graphics(fig_path_rt_ehi_cor_wide)
```


#### Statistics

Model RT as a linear effect of field, level, and EHI (continuous):
<br>
<br>
`rt_model_ehi <- lmer( rt ~ field*level*ehi + (1 | subject) )`
<br>
<br>
```{r rt_model_ehil}
## Prepare data, so emmeans contrasts show
## Global bias as a positive number.
aah_for_rt_ehi_model <- aah_for_rt_model |>
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("LVF", "RVF")))
```


```{r rt_model_ehi_caching}
if (use_cached_model_rt_ehi == FALSE) {
  rt_model_ehi <- lmer(rt ~ field:level:ehi + field:level + field:ehi + level:ehi + field + level + ehi + (1 | subject), data = aah_for_rt_ehi_model)
  
  ## Create emmeans model object, and manually cache it.
  rt_emm_ehi <- emmeans(rt_model_ehi, ~ field * level * ehi)
  
  ## Manually cache model
  saveRDS(rt_model_ehi, here(manual_cache_dir, "rt_model_ehi.rds"))
  saveRDS(rt_emm_ehi, here(manual_cache_dir, "rt_emm_ehi.rds"))
  
} else if (use_cached_model_rt_ehi == TRUE) {
  ## Load cached model
  rt_model_ehi <- readRDS(here(manual_cache_dir, "rt_model_ehi.rds"))
  rt_emm_ehi <- readRDS(here(manual_cache_dir, "rt_emm_ehi.rds"))
}
```

```{r rt_ehi_interaction_anova, echo = F}
## Use anova() on competing models to test 3-way interaction.
interaction_stats <-
  function(model_with_interaction,
           model_with_no_interaction) {
    return(anova(model_with_interaction, model_with_no_interaction))
  }

rt_model_no_interaction <- update(rt_model_ehi, . ~ . - field:level:ehi)
interaction_anova <- interaction_stats(rt_model_ehi, rt_model_no_interaction)
interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level by ehi interaction (RT)", 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
## Is this p-valued "one-sided"?
## https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html
## https://stats.libretexts.org/Bookshelves/Applied_Statistics/Book%3A_Learning_Statistics_with_R_-_A_tutorial_for_Psychology_Students_and_other_Beginners_(Navarro)/16%3A_Factorial_ANOVA/16.05%3A_The___F___test_as_a_model_comparison
```
<br>
```{r}
## Test whether interaction slope differs from zero, with emmeans

## This code estimates the effect of EHI on RT for each field*level:
rt_emt <- emtrends(rt_model_ehi, pairwise ~ field*level, var = "ehi")
## We want to compare the effect of EHI on RT for:
## (LVF Local - LVF Global) - (RVF local - RVF Global)
## Contrast indices: 2 - 5

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast(interaction = c("consec"), var = "ehi") |> 
  summary(infer = T)|>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means global bias is stronger in LVF for right handers (as predicted by AAH), in ms per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## Report EHI effect on bias by field, with CI and p-value
rt_emt$contrasts |>
  summary(infer = T) |> 
  as_tibble() |> 
  filter(contrast %in%
           c("LVF Local - LVF Global", "RVF Local - RVF Global")) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means more global bias for right handers, in ms per EHI unit (-100 to 100)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))

```
<br>
```{r}
## Report EHI effect by field and level, with CI and p-value
rt_emt$emtrends |> summary(infer = T) |>
  as_tibble() |>
  format_p.value() |>
  pretty_table() |>
  tab_header(title = "Slope of EHI and RT by field and level",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means slower RTs for right handers, in ms per EHI unit (-100 to 100).",
               locations = cells_column_labels(columns = ehi.trend)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

---

<br>

```{r}
## LVF Global bias by EHI.
## Pull estimates for strong left and right handers (ehi = -100, +100)
## Estimate effect of ehi on field x level.

## Use ref_grid (following quant 2 HW5)
ref <- ref_grid(rt_model_ehi, at = list(ehi = c(-100, 0, 100)))

## Show estimates for left and right handers.
ref |>
  contrast(interaction = c("consec"), by = "ehi") |>
  summary(infer = T, type = "response") |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Estimated LVF global bias, by EHI score") |> 
  tab_footnote(footnote = "Strong left hander: -100; Mixed hander: 0; Strong right hander: +100",
               locations = cells_column_labels(columns = ehi)) |> 
  tab_footnote(footnote = "Estimated LVF global bias (ms); a positive number means LVF global bias",
               locations = cells_column_labels(columns = estimate))
```
<br>
```{r}
## Global bias by field, by handedness.
ref |>
  contrast("pairwise", by = "ehi") |> 
  summary(infer = T, type = "response") |> 
  as_tibble() |> 
  slice(2, 5, 8, 11, 14, 17) |> 
  format_p.value() |> 
  arrange(ehi) |> 
  pretty_table() |> 
  tab_header(title = "Estimated global bias by field, by EHI score") |> 
  tab_footnote(footnote = "Estimated global bias (ms); a positive number means global bias",
               locations = cells_column_labels(columns = estimate))
```
<br>
```{r}
## Accuracy by field and level, by handedness.
ref |>
  summary(infer = T, type = "response") |> 
  as_tibble() |> 
  select(field, level, ehi, prediction, ends_with("CL")) |> 
  pretty_table() |> 
  tab_header(title = "Reaction time by field and level, by EHI score") |> 
  tab_footnote(footnote = "Reaction time (ms)",
               locations = cells_column_labels(columns = prediction))
```
<br>
$$
28.144 - 14.822 = 13.322ms \\
13.322/200 = 0.067ms / EHI unit
$$
The model estimates that an average strong right hander (EHI +100) will have **13.32ms** more LVF global bias than a strong left hander (EHI -100). Each unit change in EHI (-100:100) corresponds to a **0.067ms** difference in LVF global bias. This is also the slope estimate given by the summary function:
<br>

```{r, echo = T, results = "markup"}
summary(rt_model_ehi)
```
<br>

### Accuracy
```{r}
acc_ehi_1 <- acc_1 |>
  left_join(aah_summary |> select(subject, ehi)) |>
  select(subject, ehi, LVF_Global_Bias)
```


#### Plots
```{r}
fig_path_acc_ehi_cor <- here(fig_dir, "acc_ehi_cor.png")
fig_path_acc_ehi_cor_wide <- here(fig_dir, "acc_ehi_cor_wide.png")
g <- acc_ehi_1 |> ggplot(aes(x = ehi, y = LVF_Global_Bias)) +
  geom_beeswarm(alpha = .2) +
  geom_smooth(method = "lm", color = "black") +
  labs(x = "EHI", y = "RVF - LVF, Local - Global % correct",
       title = "EHI and each subject's mean LVF global bias")

g <- g |> gg_style()
ggsave(fig_path_acc_ehi_cor, g, "png", height = 4, width = 4)
ggsave(fig_path_acc_ehi_cor_wide, g, "png", height = 4, width = 8)
```
```{r}
include_graphics(fig_path_acc_ehi_cor_wide)
```

#### Statistics

Model accuracy as a binomial effect of field, level, and EHI (continuous):
<br>
<br>
`acc_ehi_model <- glmer( correct ~ field*level*ehi + (1 | subject), family = "binomial" )`
<br>
<br>
```{r acc_model_ehil}
## Prepare data, so emmeans contrasts show
## Global bias as a positive number.
aah_for_acc_ehi_model <- aah_for_acc_model |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("LVF", "RVF")))
```

```{r acc_model_ehi2}
## Make a binomial logistic model using data from every trial.

##TODO: To make the model converge, try:
## - NAGQ = 0 (https://stackoverflow.com/questions/44677487/lme4glmer-vs-statas-melogit-command)
## - using a different optimizer

if (use_cached_model_acc_ehi == FALSE) {
  
  ## Create and cache binomial model
  acc_model_ehi <-
    glmer(correct ~ field * level * ehi + (1 | subject),
          data = aah_for_acc_ehi_model,
          family = "binomial")
  #  Warning messages:
  #1: In checkConv(attr(opt, "derivs"), opt$par, ctrl = #control$checkConv,  :
  #  Model failed to converge with max|grad| = 0.00445801 (tol = 0.002, component 1)
  #2: In checkConv(attr(opt, "derivs"), opt$par, ctrl = #control$checkConv,  :
  #  Model is nearly unidentifiable: very large eigenvalue
  # - Rescale variables?
  
  ## I re-ran the model with the bobyqa estimate on Jan 7, 2023
  ## This gave virtually the same results
  ## (and still gave the warning about very large eigenvalue)
  # acc_model_ehi <-
  #   glmer(correct ~ field * level * ehi + (1 | subject),
  #         data = aah_for_acc_ehi_model,
  #         family = "binomial",
  #         control = glmerControl(optimizer = c("bobyqa"))
  #         )
  
  ## To address large eigenvalue error, try rescaling continuous variables.
  ## continuous variables: EHI.
  # df <- as.data.frame(aah_for_acc_ehi_model)
  # numcols <- grep("ehi|rt", names(df))
  # df_scaled <- df
  # df_scaled[, numcols] <- scale(df_scaled[, numcols])
  # acc_model_ehi_scaled <- update(acc_model_ehi, data = df_scaled)
  # acc_model_ehi_scaled
  # summary(acc_model_ehi_scaled)
  # 
  # saveRDS(acc_model_ehi_scaled, here(manual_cache_dir,
  #                             "acc_model_ehi_scaled.rds"))
  
  
  ## Create emmeans model object, and manually cache it.
  acc_emm_ehi <-
    emmeans(acc_model_ehi, ~ field * level * ehi)
  
  ## Manually cache model
  saveRDS(acc_model_ehi, here(manual_cache_dir,
                              "acc_model_ehi.rds"))
  saveRDS(acc_emm_ehi, here(manual_cache_dir, "acc_emm_ehi.rds"))
  
} else if (use_cached_model_acc_ehi == TRUE) {
  ## Load cached model
  acc_model_ehi <- readRDS(here(manual_cache_dir, "acc_model_ehi.rds"))
  acc_emm_ehi <- readRDS(here(manual_cache_dir, "acc_emm_ehi.rds"))
}
```
<br>
```{r acc_interaction_anova_ehi}
## Use anova() on competing models to test 2-way interaction.

if (use_cached_model_acc_ehi == FALSE) {
  ## Create and manually cache ANOVA
  acc_model_no_interaction <-
    update(acc_model_ehi, . ~ . - field:level:ehi)
  
  #  Warning message:
  #In checkConv(attr(opt, "derivs"), opt$par, ctrl = #control$checkConv,  :
  #  Model is nearly unidentifiable: very large eigenvalue
  # - Rescale variables?
  
  interaction_anova <-
    interaction_stats(acc_model_ehi, acc_model_no_interaction)
  
  ## Manually cache model
  saveRDS(interaction_anova,
          here(manual_cache_dir,
               "acc_anova_ehi.rds"))
  
  ## Do anova on scaled model
  # acc_model_no_interaction_scaled <-
  #   update(acc_model_ehi_scaled, . ~ . - field:level:ehi)
  # interaction_anova_scaled <-
  #   interaction_stats(acc_model_ehi_scaled, acc_model_no_interaction_scaled)
  # saveRDS(interaction_anova_scaled,
  #         here(manual_cache_dir,
  #              "acc_anova_ehi_scaled.rds"))
  
} else if (use_cached_model_acc_ehi == TRUE) {
  ## Load cached ANOVA
  interaction_anova <- readRDS(here(manual_cache_dir, "acc_anova_ehi.rds"))
}

interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level by EHI interaction (Accuracy)", 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
```
<br>
```{r}
## Test whether interaction slope differs from zero, with emmeans

## This code estimates the effect of EHI on acc for each field*level:
acc_emt <- emtrends(acc_model_ehi, pairwise ~ field*level, var = "ehi")
## We want to compare the effect of EHI on acc for:
## (LVF Global - LVF Local) - (RVF Global - RVF Local)

## Report interaction estimate, with CI and p value.
acc_emt[[1]] |> contrast(interaction = c("consec"), var = "ehi") |> 
  summary(infer = T)|>
  format_p.value() |> 
  pretty_table(digits = 4) |> 
  tab_header(title = "Field by level by EHI interaction (Accuracy)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means global bias is stronger in LVF for right handers (as predicted), in logodds per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95% (two-sided)",
               locations = cells_column_labels(columns = ends_with("CL")))
```
<br>
```{r}
## Report EHI effect on bias by field, with CI and p-value
acc_emt$contrasts |>
  summary(infer = T, adj = "none") |> 
  as_tibble() |> 
  filter(contrast %in%
           c("LVF Global - LVF Local", "RVF Global - RVF Local")) |> 
  format_p.value() |> 
  pretty_table(digits = 4) |> 
  tab_header(title = "Field by level interaction (Accuracy)") |>
  tab_footnote(footnote = "A positive number means more global bias, in logodds per EHI unit (-100 to 100)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## Report EHI effect by field and level, with CI and p-value
acc_emt$emtrends |> summary(infer = T, adj = "none") |>
  as_tibble() |>
  format_p.value() |>
  pretty_table(digits = 4) |>
  tab_header(title = "Slope of EHI and Accuracy by field and level",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means higher accuracy for right handers, in logodds per EHI unit (-100 to 100).",
               locations = cells_column_labels(columns = ehi.trend)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

---

<br>
```{r}
## LVF Global bias by handedness.
## Pull estimates for strong left and right handers (ehi = -100, +100)
## Estimate effect of ehi on field x level.

## Use ref_grid (following quant 2 HW5)
ref <- ref_grid(acc_model_ehi, at = list(ehi = c(-100, 0, 100)))


## Show estimates for left and right handers.
ref |>
  contrast(interaction = c("consec"), by = "ehi") |>
  summary(infer = T, type = "response") |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Estimated LVF global bias, by EHI score") |> 
  tab_footnote(footnote = "Strong left hander: -100; Mixed hander: 0; Strong right hander: +100",
               locations = cells_column_labels(columns = ehi)) |> 
  tab_footnote(footnote = "Estimated LVF global bias. An odds ratio > 1 means LVF global bias.",
               locations = cells_column_labels(columns = odds.ratio))
```
<br>
```{r}
## Global bias by field, by handedness.
ref |>
  contrast("pairwise", by = "ehi") |> 
  summary(infer = T, type = "response") |> 
  as_tibble() |> 
  slice(2, 5, 8, 11, 14, 17) |> 
  format_p.value() |> 
  arrange(ehi) |> 
  pretty_table() |> 
  tab_header(title = "Estimated global bias by field, by EHI score") |> 
  tab_footnote(footnote = "Estimated global bias. An odds ratio > 1 means global bias.",
               locations = cells_column_labels(columns = odds.ratio))
```
<br>
```{r}
## Accuracy by field and level, by handedness.
ref |>
  summary(infer = T, type = "response") |> 
  as_tibble() |> 
  select(field, level, ehi, prob, ends_with("CL")) |> 
  pretty_table() |> 
  tab_header(title = "Accurracy by field and level, by EHI score") |> 
  tab_footnote(footnote = "Accuracy (% correct)",
               locations = cells_column_labels(columns = prob))
```
<br>
$$
log(1.73) = .548
$$
$$
log(1.67) = .513
$$
$$
log(1.73) - log(1.67) = .0353
$$
$$
.0353 / 200 = -0.000185 logodds / EHI unit
$$
The model estimates that a strong right hander (EHI +100) will have 1.73/1.67 = 1.04 greater odds of correctness for LVF global stimuli versus a strong left hander (EHI -100). Each unit change in EHI (-100:100) corresponds to a **0.0002** (logodds) difference in LVF global bias. This matches the slope estimate given by the summary function:
<br>

```{r, echo = T, results = "markup"}
summary(acc_model_ehi)
```
<br>
Using the bobyqa optimizer alone gives virtually the same results.
```{r}
acc_model_ehi_bobyqa <- readRDS(here(manual_cache_dir, "models_bobyqa", 
                                     "acc_model_ehi.rds"))
```
```{r, echo = T, results = "markup"}
summary(acc_model_ehi_bobyqa)
```
<br>
Using a model with scaled EHI gives virtually the same results, and no warning about large eigenvalues.
```{r}
acc_model_ehi_scaled <- readRDS(here(manual_cache_dir, "models_scaled", 
                                     "acc_model_ehi_scaled.rds"))
```
```{r, echo = T, results = "markup"}
summary(acc_model_ehi_scaled)
```
<br>
