==> analysis/analyses/experiment_n1008/3e_analyze_shape_FLH_extremes.Rmd <==
---
title: "Action Asymmetry Experiment (n = 1008): Shape Analyses (EHI +/-100 only)"
pagetitle: "aah | exp | analysis (shape)"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

<style type="text/css">
  body{
  font-family: Avenir;
  font-size: 12pt;
}
</style>

---
  
# Summary

In the full sample, we saw a marginal effect of shape: right handers' vs. left handers' LVF global bias is 15.89ms greater for square targets than for circles (95% CI [-5.99, 37.76], p = .08, one-sided). We also saw that the estimated LVF global bias difference was greater between strong right and left handers (EHI +/- 100: 23.51ms difference) than between the default handedness groups (EHI </> +-40: 11.67ms difference).

**When we limit analysis to squares only and EHI extremes, strong left handers have a further reduced (very close to zero), but not reversed, LVF global bias.**

Strong left handers’ estimated LVF global bias for squares (2.45ms, 95%CI [-15.18, 20.01]) is lower than for circles (7.20ms, 95%CI [-9.98, 24.37]), but this difference is not significant (-4.74ms, 95%CI [-29.10, 19.62], p = .35, one-sided). With squares and circles combined, strong left handers' LVF global bias was 4.59ms (95%CI [-7.90, 16.98]).

The critical three-way interaction effect is stronger in this analysis of squares only and EHI extremes (34.68ms, 95%CI [11.35, 58.01], p = .002, one-sided). The three-way effect for the full sample with both shapes was 11.67ms (95%CI [0.65, 22.69], p = .019, one-sided), for the full sample with squares only it was 19.76ms (95%CI [4.11, 35.40], p = .006, one-sided), and for EHI extremes with both shapes it was 23.51ms (95% CI [7.10, 39.92], p = .0025, one-sided).


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.align = "center", fig.width = 9,
                      fig.height = 6, results = "asis")
options(knitr.kable.NA = "")

cli.progress_show_after <- 0

## Do not use scientific notation until 9 decimal places.
options(scipen = 9, digits = 9)

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)
```

```{r lib}
library(here)
library(tidyverse)
library(cli) # For printing error messages

library(lme4)
library(emmeans)
library(broom)
library(gt)

library(ggpubr) ## functions for plotting SD, SEM, CI.
library(ggbeeswarm)
library(patchwork)
library(ggh4x) ## for nested facets

library(knitr) # For include_graphics

source(here::here("lib", "load_process", "load_process.R"))
source(here::here("lib", "util.R"))
source(here::here("lib", "demographics.R"))
source(here::here("lib", "plots.R"))
source(here::here("lib", "3c_analyze_shape", "stats.R"))

## Colors for plots
plot_blue <- c("#337aB7")
plot_colors <- c("#4477AA",
                 "#EE6677",
                 "#CCBB44",
                 "#66CCEE",
                 "#AA3377",
                 "#228833",
                 "#BBBBBB")
```

```{r config}
proc_dir <- here::here("data", "proc_exp_n1008")
fig_dir <- here::here("figures", "exp_n1008_e")
manual_cache_dir <- here::here("manual_cache", "exp_n1008_e")

use_cached_models <- T

use_cached_figs <- T

if (use_cached_figs == TRUE) {
  use_cached_demofigs <- TRUE
  use_cached_rtfigs <- TRUE
} else if (use_cached_figs == FALSE) {
  use_cached_demofigs <- FALSE
  use_cached_rtfigs <- FALSE
}
```

---

# Analyses {.tabset}
```{r load_data}
## Load "the data" with all subjects & trials. Filter to only subjects with EHI -100 or +100.
aah_long <- load_aah_long(proc_dir) |> mutate(
  handedness = case_when(
    ehi == -100 ~ "Left",
    ehi == 100 ~ "Right",
    ehi > -100 & ehi < 100 ~ "Mixed")
) |> filter(handedness %in% c("Right", "Left"))

## Load summary data table (for quick demographic analyses)
## with all subjects.
aah_summary_all <- load_aah_summary(proc_dir) |> mutate(
  handedness = case_when(
    ehi == -100 ~ "Left",
    ehi == 100 ~ "Right",
    ehi > -100 & ehi < 100 ~ "Mixed")
) |> filter(handedness %in% c("Right", "Left"))
```

```{r prepare_data}
#### PREPARE SUMMARY DATA FOR DEMOGRAPHICS ANALYSES
aah_summary <- filter_aah_summary_for_analysis(aah_summary_all)

#### PREPARE TRIAL-LEVEL DATA FOR RT ANALYSES
## Prepare "the data" (aah_long) for all analyses:
## Filter out practice trials, absent trials, and excluded subjects
## This data (all present trials, correct or incorrect) will be used for
## accuracy analyses
aah <- filter_aah_long_for_analysis(aah_long) |> 
  rename(shape = target)

## For RT analyses, prepare dataset with only correct, present trials.
## In our RT model, we only care about correct responses to present trials.
aah_correct <- aah |> filter(correct == T)
## Code handedness

## Relevel field and level with RVF first (unintuitive for plotting),
## so that emmeans will show a positive number for LVF global bias.
aah_for_rt_model <- aah_correct |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("RVF", "LVF")))


#### PREPARE SUBJECT_LEVEL DATA FOR RT ANALYSES
## Prepare subject-level data (RT)
rt_subject <- aah_correct |> group_by(subject, field, level, handedness, shape) |>
  summarize(rt = median(rt))

## Prepare subject-level LVF Global bias summary (RT)
rt_1 <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  mutate(all_one_group = "all_one_group")

```

## Field by level by shape {.tabset .tabset-pills}

Does the interaction of field by level differ depending on shape (square vs. circle), for extreme left and right handers (EHI +/- 100)?

Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*shape + (1 | subject) )`
<br>
<br>

### Extreme righties (EHI +100; n = 182)
<br>
```{r FLS_demo_182RH}
data_label <- "182RH"
data_sub <- aah_for_rt_model |> filter(handedness == "Right")
data_sub_summary <- aah_summary |> filter(handedness == "Right")

FLS_demo <- demo_summary_table(data_sub_summary)
FLS_demo |> pretty_table() |> tab_header(title = "Demographics")
```
<br>
```{r FLS_model_emm_182RH}
FLS <- model_emm_FLS(
  use_cached_model = use_cached_models,
  manual_cache_dir = manual_cache_dir,
  data = data_sub,
  data_label = data_label
)
```
<br>
```{r FLS_anova_182RH}
## ANOVA: is there a significant interaction of field by level by shape?
anova_FLS(use_cached_anova = use_cached_models,
          manual_cache_dir = manual_cache_dir,
          data_label = data_label,
          FLS$model,
          title = "Field by level by shape interaction (RT)")
```
<br>
```{r FLS_emmtest_182RH}
## Use emmeans() to test 3-way interaction.
emmtest_FLS(FLS$emm, title = "Field by  level by shape interaction (RT)")
```
<br>
```{r FL_S_emmtest_182RH}
## Estimate interaction of field by level for  shape
emmtest_FL_S(use_cached_model = use_cached_models,
             manual_cache_dir = manual_cache_dir,
             FLS$model,
             data_label = data_label,
             title = "Effect of field by level for each shape (RT)")
```
<br>
```{r FG_F_S_182RH}
## Estimate global bias by field, for each shape
emmtest_GB_F_S(FLS$emm, data_label, title = "Global bias by field, by shape (RT)")
```
<br>
```{r F_L_S_emmtest_182RH}
emmtest_F_L_S(FLS$emm, data_label, title = "RT estimates by field, level, and shape")
```
<br>

### Extreme lefties (EHI -100; n = 182)
<br>
```{r FLS_demo_138LH}
data_label <- "138LH"
data_sub <- aah_for_rt_model |> filter(handedness == "Left")
data_sub_summary <- aah_summary |> filter(handedness == "Left")

FLS_demo <- demo_summary_table(data_sub_summary)
FLS_demo |> pretty_table() |> tab_header(title = "Demographics")
```
<br>
```{r FLS_model_emm_138LH}
FLS <- model_emm_FLS(
  use_cached_model = use_cached_models,
  manual_cache_dir = manual_cache_dir,
  data = data_sub,
  data_label = data_label
)
```
<br>
```{r FLS_anova_138LH}
## ANOVA: is there a significant interaction of field by level by shape?
anova_FLS(use_cached_anova = use_cached_models,
          manual_cache_dir = manual_cache_dir,
          data_label = data_label,
          FLS$model,
          title = "Field by level by shape interaction (RT)")
```
<br>
```{r FLS_emmtest_138LH}
## Use emmeans() to test 3-way interaction.
emmtest_FLS(FLS$emm, title = "Field by  level by shape interaction (RT)")
```
<br>
```{r FL_S_emmtest_138LH}
## Estimate interaction of field by level for  shape
emmtest_FL_S(use_cached_model = use_cached_models,
             manual_cache_dir = manual_cache_dir,
             FLS$model,
             data_label = data_label,
             title = "Effect of field by level for each shape (RT)")
```
<br>
```{r FG_F_S_138LH}
## Estimate global bias by field, for each shape
emmtest_GB_F_S(FLS$emm, data_label, title = "Global bias by field, by shape (RT)")
```
<br>
```{r F_L_S_emmtest_138LH}
emmtest_F_L_S(FLS$emm, data_label, title = "RT estimates by field, level, and shape")
```
<br>

## Field by level by handedness by shape {.tabset .tabset-pills}

Does the interaction of field by level by handedness differ depending on shape (square vs. circle), for extreme left and right handers (EHI +/- 100)?

Reaction time is modeled as a linear effect of field, level, handedness (EHI < -100 or EHI > +100), and shape, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field * level * handedness * shape + (1 | subject) )`
<br>
<br>
```{r FLSH_demo_2bins}
data_label <- "2bins"
data_sub <- aah_for_rt_model |> filter(handedness %in% c("Right", "Left"))
data_sub_summary <- aah_summary |> filter(handedness %in% c("Right", "Left"))

FLS_demo <- hand_demo_summary_table(data_sub_summary)
FLS_demo |> pretty_table() |> tab_header(title = "Demographics")
```
<br>
```{r FLSH_model_emm_2bins}
FLSH <- model_emm_FLSH(
  use_cached_model = use_cached_models,
  manual_cache_dir = manual_cache_dir,
  data = data_sub,
  data_label = data_label
)
```
<br>
```{r FLSH_anova_2bins}
## ANOVA: is there a significant interaction of field by level by shape?
anova_FLSH(use_cached_anova = use_cached_models,
          manual_cache_dir = manual_cache_dir,
          data_label = data_label,
          FLSH$model,
          title = "4-way field by level by handedness by shape interaction (RT)")
```
<br>
```{r FLSH_emmtest_2bins}
## Use emmeans() to test 4-way interaction.
emmtest_FLSH(FLSH$emm, title = "4-way field by level by handedness by shape interaction (RT)")
```
<br>
```{r FLH_S_emmtest_2bins}
## Estimate the 3-way interaction of field by level by handedness, for each shape
emmtest_FLH_S(use_cached_model = use_cached_models,
             manual_cache_dir = manual_cache_dir,
             FLSH$model,
             data_label = data_label,
             title = "Effect of field by level by handedness for each shape (RT)")
```
<br>
```{r FLS_H_emmtest_2bins}
## Estimate the 3-way interaction of field by level by shape, for each handedness group
emmtest_FLS_H(use_cached_model = use_cached_models,
             manual_cache_dir = manual_cache_dir,
             FLSH$model,
             data_label = data_label,
             title = "Effect of field by level by shape for each handedness group (RT)")
```
<br>
```{r FL_H_S_2bins}
## Estimate the field by level interaction, for each handedness group, for each shape.
emmtest_FL_H_S(use_cached_model = use_cached_models,
               manual_cache_dir = manual_cache_dir,
               FLSH$emm,
               data_label = data_label,
               title = "Effect of field by level for each handedness group, by shape (RT)")
```
<br>


## Plots {.active}

---

Error bars show 95% CI. Left handers are those with EHI = -100; Right handers, EHI = +100.


```{r plot_rt_config}
plot_color <- plot_colors[[5]] # Color for interaction plots with only one color.

handedness_labels <- c("Strong right handers (EHI +100)", "Strong left handers (EHI -100)")
names(handedness_labels) <- c("Right", "Left")

handedness_labels_short <- c("Right (EHI +100)", "Left (EHI -100)")
names(handedness_labels_short) <- c("Right", "Left")
```

---

```{r plot_rt_1}
fig_path_rt_1 <- here(fig_dir, "rt_1.png")
fig_path_rt_1_wide <- here(fig_dir, "rt_1_wide.png")

if (use_cached_rtfigs == FALSE) {
  ## Make a table showing:
  ## For each subject and field, the difference in median rt for:
  ## Global - Local
  rt_1 <- rt_subject |>
    pivot_wider(names_from = c(field, level),
                values_from = rt) |>
    mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
    mutate(all_one_group = "all_one_group")
    
  rt_descriptive <- rt_1 |>
    group_by(all_one_group, handedness, shape) |> 
    summarize(
      median = median(LVF_Global_Bias),
      mean = mean(LVF_Global_Bias),
      SE = sd(LVF_Global_Bias) / sqrt(length((LVF_Global_Bias)))
    )
  
  g <- ggplot(rt_1, aes(
    x = all_one_group,
    y = LVF_Global_Bias,
    fill = handedness,
    color = handedness
  )) +
    geom_hline(yintercept = 0, color = "gray50", linewidth = .5) +
    geom_quasirandom(
      alpha = .2, show.legend = F,
      ) +
    geom_boxplot(
      alpha = 0.5,
      size = 0.3,
      varwidth = F,
      outlier.shape = NA,
      color = "gray20",
      # fill = plot_color,
      show.legend = F
    ) +
    stat_summary(
      fun.data = mean_cl_normal,
      fun.args = list(conf.int = 0.95),
      geom = "errorbar",
      position = "dodge",
      linetype = 1,
      color = "gray5",
      width = .2,
      # size = point_size / 5,
      show.legend = F
    ) +
    geom_point(
      data = rt_descriptive,
      aes(y = mean),
      # fill = plot_color,
      color = "black",
      shape = 21,
      show.legend = F
    ) +
    ggh4x::facet_nested(~ shape + handedness,
               labeller = labeller(handedness = handedness_labels_short),
               nest_line = element_line()) +
    scale_y_continuous(minor_breaks = seq(-1000 , 1000, 50),
                       breaks = seq(-1000, 1000, 100)) +
    scale_fill_manual(values = plot_colors[c(1, 5,  2)]) +
    scale_color_manual(values = plot_colors[c(1, 5,  2)]) +
    theme(axis.text.x = element_blank()) +
    labs(title = "LVF>RVF Global Bias by target shape", x = "Level", y = "RVF - LVF, Local - Global RT (ms)")
  
  g <- g |> gg_style_means() +
    theme(axis.text.x = element_blank(), aspect.ratio = 4/1,
          strip.background = element_blank(),
          ggh4x.facet.nestline = element_line(color = "gray50"))
  
  ggsave(fig_path_rt_1, g, "png", height = 5, width = 5)
  ggsave(fig_path_rt_1_wide, g, "png", height = 5, width = 8)
  g_rt_1 <- g
}

include_graphics(fig_path_rt_1_wide)
```

---

```{r plot_rt_2}
fig_path_rt_2 <- here(fig_dir, "rt_2.png")

if (use_cached_rtfigs == FALSE) {
## Make a table showing:
## For each subject and field, the difference in median rt for:
## Global - Local
rt_2 <- rt_subject |> 
  pivot_wider(names_from = c(level),
              values_from = rt) |> 
  mutate(Global_Bias = Local - Global)

rt_descriptive <- rt_2 |>
  group_by(field, handedness, shape) |>
  summarize(
    median = median(Global_Bias),
    mean = mean(Global_Bias),
    SE = sd(Global_Bias) / sqrt(length((Global_Bias)))
  )

g <- ggplot(rt_2, aes(x = field,
                      y = Global_Bias, fill = handedness, color = handedness)) +
  geom_hline(yintercept = 0, color = "gray50", linewidth = .5) +
  geom_quasirandom(
    alpha = .2, show.legend = F,
    # color = plot_color
    ) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    # fill = plot_color,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .2,
    show.legend = F
  ) +
  geom_point(
    data = rt_descriptive,
    aes(y = mean),
    # fill = plot_color,
    color = "gray5",
    shape = 21,
    show.legend = F
  ) +
  ggh4x::facet_nested(~ shape + handedness,
             labeller = labeller(handedness = handedness_labels_short),
             nest_line = element_line()) +
  scale_y_continuous(minor_breaks = seq(-1000 , 1000, 50),
                     breaks = seq(-1000, 1000, 100)) +
  scale_fill_manual(values = plot_colors[c(1, 5,  2)]) +
  scale_color_manual(values = plot_colors[c(1, 5,  2)]) +
  labs(title = "Global bias by target shape", x = "Level", y = "Local - Global RT (ms)")

g <- g |>  gg_style_means() |> gg_color() +
  theme(strip.background = element_blank(),
        ggh4x.facet.nestline = element_line(color = "gray50"))

ggsave(fig_path_rt_2, g, "png", height = 4, width = 6)
g_rt_2 <- g
}
include_graphics(fig_path_rt_2)
```

---

```{r plot_rt_4}
fig_path_rt_4 <- here(fig_dir, "rt_4.png")

if (use_cached_rtfigs == FALSE) {
  rt_subject <- aah_correct |>
    group_by(subject, field, level, handedness, shape) |> 
    summarize(rt = median(rt))
  
  rt_descriptive <- rt_subject |>
    group_by(field, level, handedness, shape) |>
    summarize(
      median = median(rt),
      mean = mean(rt),
      SE = sd(rt) / sqrt(length((rt)))
    )
  
  g <- ggplot(rt_subject, aes(
    x = level,
    y = rt,
    fill = level,
    color = level
  )) +
    geom_hline(yintercept = 0, color = "gray50", linewidth = .5) +
    geom_quasirandom(alpha = .3, show.legend = F) +
    geom_boxplot(
      alpha = 0.5,
      size = 0.3,
      varwidth = F,
      outlier.shape = NA,
      color = "gray20",
      show.legend = F
    ) +
    stat_summary(
      fun.data = mean_cl_normal,
      fun.args = list(conf.int = 0.95),
      geom = "errorbar",
      position = "dodge",
      linetype = 1,
      color = "gray5",
      width = .2,
      # size = point_size / 5,
      show.legend = F
    ) +
    geom_point(
      data = rt_descriptive,
      aes(y = mean),
      color = "black",
      shape = 21,
      show.legend = F
    ) +
    scale_y_continuous(
      minor_breaks = seq(0 , 1500, 100),
                       breaks = seq(0, 1500, 200),
      expand = expansion(mult = c(0, .08)), limits = c(0, NA)) +
    ggh4x::facet_nested(shape ~ handedness + field,
               labeller = labeller(handedness = handedness_labels),
               nest_line = element_line()) +
    labs(title = "Per-subject medians (RT) by target shape", x = "Level", y = "Reaction time (ms)")
  
  g <- g |>  gg_style_means() |> gg_color() +
    theme(strip.background = element_blank(),
          ggh4x.facet.nestline = element_line(color = "gray50"))
      
  ggsave(fig_path_rt_4, g, "png", height = 6, width = 6)
  g_rt_4 <- g
}

include_graphics(fig_path_rt_4)
```

---

```{r plot_rt_4a_all_trials}
## Plot data from every trial, with overall mean, median and bounds.
fig_path_rt_4a <- here(fig_dir, "rt_4a.png")

if (use_cached_rtfigs == FALSE) {
rt_descriptive <- aah_correct |>
  group_by(field, level, handedness, shape) |>
  summarize(
    median = median(rt),
    mean = mean(rt),
    SE = sd(rt) / sqrt(length((rt)))
  )

## Facet by: Visual field, handedness
g <- ggplot(aah_correct, aes(
  x = level,
  y = rt,
  fill = level,
  color = level,
)) +
  geom_quasirandom(alpha = .05, show.legend = F) +
  geom_boxplot(
    alpha = 0.5,
    size = 0.3,
    varwidth = F,
    outlier.shape = NA,
    color = "gray20",
    show.legend = F
  ) +
  stat_summary(
    fun.data = mean_cl_normal,
    fun.args = list(conf.int = 0.95),
    geom = "errorbar",
    position = "dodge",
    linetype = 1,
    color = "gray5",
    width = .3,
    # size = point_size / 5,
    show.legend = F
  ) +
  geom_point(data = rt_descriptive, aes(y = mean), color = "black", shape = 21,
             show.legend = F) +
  scale_y_continuous(minor_breaks = seq(0 , 2000, 100),
                     breaks = seq(0, 2000, 200)) +
  ggh4x::facet_nested(shape ~ handedness + field,
             labeller = labeller(handedness = handedness_labels),
             nest_line = element_line()) +
  labs(title = "All trials (RT) by target shape", x = "Level", y = "Reaction time (ms)")

g <- g |>  gg_style_means() |> gg_color() +
  theme(strip.background = element_blank(),
        ggh4x.facet.nestline = element_line(color = "gray50"))

ggsave(fig_path_rt_4a, g, "png", height = 6, width = 6)
g_rt_4a <- g
}

include_graphics(fig_path_rt_4a)
```
<br>
<br>
