==> analysis/analyses/experiment_n1008/4a_figures_cogsci2023.Rmd <==
---
title: "Action Asymmetry Experiment (n = 1008): Figures for CogSci 2023 submission"
pagetitle: "aah | exp | figures"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

<style type="text/css">
  body{
  font-family: Avenir;
  font-size: 12pt;
}
</style>

---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.align = "center", fig.width = 9,
                      fig.height = 6, results = "asis")
options(knitr.kable.NA = "")

cli.progress_show_after <- 0

## Do not use scientific notation until 9 decimal places.
options(scipen = 9, digits = 9)

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)
```

```{r lib}
library(here)
library(tidyverse)
library(cli) # For printing error messages

library(lme4)
library(emmeans)
library(broom)
library(gt)

library(ggpubr) ## functions for plotting SD, SEM, CI.
library(ggbeeswarm)
library(patchwork)
library(ggh4x) ## for nested facets

library(knitr) # For include_graphics

source(here::here("lib", "load_process", "load_process.R"))
source(here::here("lib", "util.R"))
source(here::here("lib", "demographics.R"))
source(here::here("lib", "plots.R"))
source(here::here("lib", "3c_analyze_shape", "stats.R"))

## Colors for plots
plot_blue <- c("#337aB7")
plot_colors <- c("#4477AA",
                 "#EE6677",
                 "#CCBB44",
                 "#66CCEE",
                 "#AA3377",
                 "#228833",
                 "#BBBBBB")
```

```{r config}
proc_dir <- here::here("data", "proc_exp_n1008")
fig_dir <- here::here("figures", "exp_n1008_4a")
manual_cache_dir <- here::here("manual_cache", "exp_n1008_4a")

use_cached_models <- T

use_cached_figs <- F

if (use_cached_figs == TRUE) {
  use_cached_demofigs <- TRUE
  use_cached_rtfigs <- TRUE
} else if (use_cached_figs == FALSE) {
  use_cached_demofigs <- FALSE
  use_cached_rtfigs <- FALSE
}
```

```{r load_data}
## Load "the data" with all subjects & trials.
aah_long <- load_aah_long(proc_dir)

## Load summary data table (for quick demographic analyses)
## with all subjects.
aah_summary_all <- load_aah_summary(proc_dir)
```

```{r prepare_data}
#### PREPARE SUMMARY DATA FOR DEMOGRAPHICS ANALYSES
aah_summary <- filter_aah_summary_for_analysis(aah_summary_all) |>
  mutate(
    handedness_extremes = case_when(ehi == -100 ~ "Left",
                                    ehi == 100 ~ "Right",
                                    ehi > -100 & ehi < 100 ~ "Mixed")
  )


#### PREPARE TRIAL-LEVEL DATA FOR RT, ACCURACY ANALYSES
## Prepare "the data" (aah_long) for all analyses:
## Filter out practice trials, absent trials, and excluded subjects
## This data (all present trials, correct or incorrect) will be used for
## accuracy analyses
aah <- filter_aah_long_for_analysis(aah_long) |>
  mutate(
    handedness_extremes = case_when(ehi == -100 ~ "Left",
                                    ehi == 100 ~ "Right",
                                    ehi > -100 &
                                      ehi < 100 ~ "Mixed")
  ) |> rename(shape = target)

## For RT analyses, prepare dataset with only correct, present trials.
## In our RT model, we only care about correct responses to present trials.
aah_correct <- aah |> filter(correct == T)
## Code handedness

## Relevel field and level with RVF first (unintuitive for plotting),
## so that emmeans will show a positive number for LVF global bias.
aah_for_rt_model <- aah_correct |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("RVF", "LVF")))

## For accuracy analyses, prepare dataset will all present trials.
## Relevel field and level,
## so that emmeans will show a positive number for LVF global bias.
aah_for_acc_model <- aah |>
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("RVF", "LVF")))

#### PREPARE SUBJECT_LEVEL DATA FOR RT, ACCURACY ANALYSES
## Prepare subject-level data (RT)
rt_subject <- aah_correct |> group_by(subject, field, level, handedness) |>
  summarize(rt = median(rt))

## Prepare subject-level LVF Global bias summary (RT)
rt_1 <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  mutate(all_one_group = "all_one_group")

## Prepare subject-level LVF Global bias summary (Accuracy)
## (For accuracy, include incorrect trials!)
acc_subject <- aah |> group_by(subject, field, level, handedness) |>
  summarize(
    total_responses = n(),
    n_present_resp = sum(correct),
    n_absent_resp = total_responses - n_present_resp,
    n_correct = sum(correct),
    acc = 100 * (n_correct / total_responses)
  ) |>
  select(subject, handedness, field, level, acc) |>
  mutate(level = recode(level, global = "Global", local = "Local"))

acc_1 <- acc_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = acc) |>
  mutate(LVF_Global_Bias = (LVF_Global - LVF_Local) - (RVF_Global - RVF_Local)) |>
  mutate(all_one_group = "all_one_group")
```

# {.tabset .tabset-pills}

## Demographics

Demographic numbers to put in tables or text.

### Everyone

Demographics for all included participants.
```{r demo_summary}
rt_demo <- demo_summary_table(aah_summary)
rt_demo |> pretty_table() |> tab_header(title = "Demographics", subtitle = "Summary")
```
<br>
```{r demo_race}
demo_race_table(aah_summary) |> pretty_table()
```
<br>
```{r demo_ethnicity}
demo_ethnicity_table(aah_summary) |> pretty_table()
```
<br>

### By handedness group

Demographics for included participants, by handedness group (EHI bins). 
```{r}
hand_demo_summary_table(aah_summary) |> pretty_table() |>
  tab_footnote("Left: (EHI <= -40)  |  Mixed: (-40 < EHI < 40)  |  Right: (EHI >= 40)")

## Save as latex object
# out_table <- hand_demo_summary_table(aah_summary) |> pretty_table() |>
#   tab_footnote("Left: (EHI <= -40)  |  Mixed: (-40 < EHI < 40)  |  Right: (EHI >= 40)")
# here("tables")
# gt::gtsave(out_table, filename = "table1.tex", path = here("tables"))
```
<br>

## EHI distribution

```{r ehi_distribution_plot}
fig_path_var <- here(fig_dir, "demo_ehi_dots_blue.png")

if (use_cached_rtfigs == F) {
  ehi_plot_data <- aah_summary |> group_by(ehi) |> summarize(n = n())
  g <- gg_ehi_dotarea(ehi_plot_data, plot_color = plot_blue)
  ggsave(fig_path_var, g, "png", height = 3, width = 6)
}

include_graphics(fig_path_var)
```
<br>
<br>
<br>
<br>
```{r ehi_distribution_plot_purple1}
middle_purple1 <- "#996F91" ## In between plot colors one and two
fig_path_var <- here(fig_dir, "demo_ehi_dots_purple.png")

if (use_cached_rtfigs == T) {
  ehi_plot_data <- aah_summary |> group_by(ehi) |> summarize(n = n())
  g <- gg_ehi_dotarea(ehi_plot_data, plot_color = middle_purple1)
  ggsave(fig_path_var, g, "png", height = 3, width = 6)
}

include_graphics(fig_path_var)
```
<br>
<br>
<br>
<br>
```{r ehi_distribution_plot_purple2}
middle_purple2 <- "#886e92" ## In between plot colors one and two
fig_path_var <- here(fig_dir, "demo_ehi_dots_purple2.png")

if (use_cached_rtfigs == T) {
  ehi_plot_data <- aah_summary |> group_by(ehi) |> summarize(n = n())
  g <- gg_ehi_dotarea(ehi_plot_data, plot_color = middle_purple2)
  ggsave(fig_path_var, g, "png", height = 3, width = 6)
}

include_graphics(fig_path_var)
```


## Preregistered 3-way (RT, bins)

Show interaction at each level: 8 boxes, 4 boxes, 2 boxes.

---

Dots show per-subject means, error bars show SEM.

```{r plot_rt_config}
plot_color <- plot_colors[[5]] # Color for interaction plots with only one color.

# handedness_labeller_40 <- c(
#   Right = "Right handers (EHI >= +40)",
#   Left = "Left handers (EHI <= -40)"
# )
# 
# handedness_labeller_40_short <- c(
#   Right = "Right \n EHI > +40",
#   Left = "Left \n EHI < -40"
# )
# 
# handedness_labeller_generic_short <- c(
#   Right = "Right \n handers",
#   Left = "Left \n handers"
# )
```

```{r plot_rt_1_horizontal}
fig_path_var <- here(fig_dir, "rt_1_horizontal.png")

if (use_cached_rtfigs == FALSE) {
  rt_subject_plot <- rt_subject |>
    filter(handedness %in% c("Left", "Right"))
  
    g <- gg_rt_1_horizontal(title = "Global bias asymmetry by handedness (EHI cut at +/-40)",
                          rt_subject_plot,  plot_colors
    )
    
  ggsave(fig_path_var, g, "png", height = 3, width = 8)
}

include_graphics(fig_path_var)
```

```{r plot_rt_2_horizontal}
fig_path_var <- here(fig_dir, "rt_2_horizontal.png")

if (use_cached_rtfigs == FALSE) {
  rt_subject_plot <- rt_subject |>
    filter(handedness %in% c("Left", "Right"))
  
    g <- gg_rt_2_horizontal(title = "Hemifield bias for global and local targets, by handedness (EHI cut at +/-40)",
                          rt_subject_plot,
                          handedness_labeller = NULL,
                          plot_colors = plot_colors
    )
    
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```

```{r plot_rt_4_horizontal}
fig_path_var <- here(fig_dir, "rt_4_horizontal.png")

if (use_cached_rtfigs == F) {
  rt_subject_plot <- rt_subject |>
    filter(handedness %in% c("Left", "Right"))
  
    g <- gg_rt_4_horizontal(title = "Reaction time by level, field, and handedness (EHI cut at +/-40)",
                          rt_subject_plot,
                          handedness_labeller = NULL,
                          plot_colors = plot_colors
    )
    
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```

---

**TODO?**. Show model-estimated mean & CI, instead of per-subject.

<br>

<!-- ## Preregistered 3-way (accuracy, bins) -->

<!-- Show interaction at each level: 8 boxes, 4 boxes, 2 boxes. -->


## Zoom: EHI +/- 100

Show 3-way (categorical) for strong left vs. right handers. Error bars show SEM.

```{r plot_rt_1_horizontal_zoom}
fig_path_var <- here(fig_dir, "rt_1_horizontal_zoom.png")

if (use_cached_rtfigs == FALSE) {
  
  rt_subject_plot <- aah_correct |>
    group_by(subject, field, level, handedness_extremes) |>
    summarize(rt = median(rt)) |>
    filter(handedness_extremes %in% c("Left", "Right")) |>
    rename(handedness = handedness_extremes) |>
    filter(handedness %in% c("Left", "Right"))
  
  g <- gg_rt_1_horizontal(title = "Global bias asymmetry by handedness (EHI = +/-100)",
                          rt_subject_plot, plot_colors
    )
  
  ggsave(fig_path_var, g, "png", height = 3, width = 8)
}

include_graphics(fig_path_var)
```

```{r plot_rt_2_horizontal_zoom}
fig_path_var <- here(fig_dir, "rt_2_horizontal_zoom.png")

if (use_cached_rtfigs == FALSE) {
  
  rt_subject_plot <- aah_correct |>
    group_by(subject, field, level, handedness_extremes) |>
    summarize(rt = median(rt)) |>
    filter(handedness_extremes %in% c("Left", "Right")) |>
    rename(handedness = handedness_extremes) |>
    filter(handedness %in% c("Left", "Right"))
  
    g <- gg_rt_2_horizontal(title = "Hemifield bias for global and local targets, by handedness (EHI = +/-100)",
                          rt_subject_plot,
                          handedness_labeller = NULL,
                          plot_colors = plot_colors
    )
    
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```

```{r plot_rt_4_horizontal_zoom}
fig_path_var <- here(fig_dir, "rt_4_horizontal_zoom.png")

if (use_cached_rtfigs == F) {
  
  rt_subject_plot <- aah_correct |>
    group_by(subject, field, level, handedness_extremes) |>
    summarize(rt = median(rt)) |>
    filter(handedness_extremes %in% c("Left", "Right")) |>
    rename(handedness = handedness_extremes) |>
    filter(handedness %in% c("Left", "Right"))
  
    g <- gg_rt_4_horizontal(title = "Reaction time by level, field, and handedness (EHI = +/-100)",
                          rt_subject_plot,
                          handedness_labeller = NULL,
                          plot_colors = plot_colors
    )
    
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```

## Zoom: Shape

Show 3-way (categorical) for strong left vs. right handers, for squares only (and/or squares vs circles). Error bars show SEM.
```{r}
if (use_cached_rtfigs == FALSE) {
  rt_subject_plot_zoom_squares <- aah_correct |>
    filter(shape == "Square") |>
    group_by(subject, field, level, handedness_extremes) |>
    summarize(rt = median(rt)) |>
    filter(handedness_extremes %in% c("Left", "Right")) |>
    rename(handedness = handedness_extremes) |>
    filter(handedness %in% c("Left", "Right"))
}
```

```{r plot_rt_1_horizontal_zoom_squares}
fig_path_var <- here(fig_dir, "rt_1_horizontal_zoom_squares.png")

if (use_cached_rtfigs == FALSE) {
  g <- gg_rt_1_horizontal(
    title = "Global bias asymmetry by handedness (EHI = +/- 100), square targets only",
    rt_subject_plot = rt_subject_plot_zoom_squares,
    plot_colors
    )
  
  ggsave(fig_path_var, g, "png", height = 3, width = 8)
}

include_graphics(fig_path_var)
```

```{r plot_rt_2_horizontal_zoom_squares}
fig_path_var <- here(fig_dir, "rt_2_horizontal_zoom_squares.png")

if (use_cached_rtfigs == F) {
  
  g <- gg_rt_2_horizontal(
      title = "Hemifield bias for global and local targets, by handedness (EHI = +/-100), squares only",
      rt_subject_plot = rt_subject_plot_zoom_squares,
      plot_colors)
    
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```

```{r plot_rt_4_horizontal_zoom_squares}
fig_path_var <- here(fig_dir, "rt_4_horizontal_zoom_squares.png")

if (use_cached_rtfigs == F) {
  
  rt_subject_plot <- aah_correct |>
    filter(shape == "Square") |> 
    group_by(subject, field, level, handedness_extremes) |>
    summarize(rt = median(rt)) |>
    filter(handedness_extremes %in% c("Left", "Right")) |>
    rename(handedness = handedness_extremes) |>
    filter(handedness %in% c("Left", "Right"))
  
    g <- gg_rt_4_horizontal(
      title = "Hemifield bias for global and local targets, by handedness (EHI = +/-100), squares only",
      rt_subject_plot = rt_subject_plot_zoom_squares,
      plot_colors
    )
    
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```


<!-- ## Zoom: EHI +/- 100, squares only -->

<!-- Show 3-way (categorical) for strong left vs. right handers, squares only. -->



## 3-way (bins), model estimates {.active}

Diamonds and lineranges show mixed-effects model point estimates and 95% CI.


```{r plot_cat_model}
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_lmer.png")
fig_path_var_wide <- here(fig_dir, "rt_1_lmer_wide.png")


model_cache_dir <- here("manual_cache", "exp_n1008_b") ## 3b: FLH
rt_emm <- readRDS(here(model_cache_dir, "rt_emm_2bins.rds"))

## Get estimates for right and left handers.
emm_data_2 <- rt_emm |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95) |>
  as_tibble() |>
  ## Fix units
  mutate(
    estimate = estimate * -1,
    asymp.LCL = asymp.LCL * -1,
    asymp.UCL = asymp.UCL * -1
  )


plot_data <- emm_data_2 |> mutate(dv = estimate)

g <- gg_rt_2_lmer(
  title = "Hemifield bias by level (EHI cut at +/-40)",
  plot_data,
  handedness_labeller = NULL,
  plot_colors = plot_colors,
  direction_labels = list(up = "LVF bias",
                          down = "RVF bias"),
  direction_labels_pos = list(up = 38,
                              down = -38),
  ylims = list(upper = 38,
               lower = -38),
  ybreaks = list(major = 10,
                 minor = 5),
  n_subjects = list(
    right = 378,
    left = 331
  )
)

ggsave(fig_path_var, g, "png", height = 3, width = 4)

include_graphics(fig_path_var)
```
<br>
```{r plot_cat_model_zoom}
## Categorical plots: EHI extremes
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_lmer_zoom.png")
fig_path_var_wide <- here(fig_dir, "rt_1_lmer_zoom_wide.png")

rt_emm <- readRDS(here(model_cache_dir, "rt_emm_2bins.rds"))

## Get estimates for right and left handers.
zoom_data <- aah_for_rt_model |>
  filter(handedness_extremes %in% c("Right", "Left"))

rt_model <- lmer(data = zoom_data,
                 rt ~ field * level * handedness + (1 |subject)
            )

rt_emm <- rt_model |> emmeans(~field * level * handedness)

emm_data_2 <- rt_emm |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95) |>
  as_tibble() |>
  ## Fix units
  mutate(
    estimate = estimate * -1,
    asymp.LCL = asymp.LCL * -1,
    asymp.UCL = asymp.UCL * -1
  )


plot_data <- emm_data_2 |> mutate(dv = estimate)

g <- gg_rt_2_lmer(
  title = "Hemifield bias by level (EHI +/-100)",
  plot_data,
  handedness_labeller = NULL,
  plot_colors = plot_colors,
  direction_labels = list(up = "LVF bias",
                          down = "RVF bias"),
  direction_labels_pos = list(up = 38,
                              down = -38),
  ylims = list(upper = 38,
               lower = -38),
  ybreaks = list(major = 10,
                 minor = 5),
  n_subjects = list(
    right = 182,
    left = 138
  )
)

ggsave(fig_path_var, g, "png", height = 3, width = 4)

include_graphics(fig_path_var)
```

<br>
```{r plot_cat_model_squares}
## Categorical plots: EHI extremes
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_lmer_squares.png")
fig_path_var_wide <- here(fig_dir, "rt_1_lmer_squares_wide.png")

rt_emm <- readRDS(here(model_cache_dir, "rt_emm_2bins.rds"))

## Get estimates for right and left handers.
squares_data <- aah_for_rt_model |>
  filter(handedness %in% c("Right", "Left")) |> 
  filter(shape == "Square")

rt_model <- lmer(data = squares_data,
                 rt ~ field * level * handedness + (1 |subject)
            )

rt_emm <- rt_model |> emmeans(~field * level * handedness)

emm_data_2 <- rt_emm |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95) |>
  as_tibble() |>
  ## Fix units
  mutate(
    estimate = estimate * -1,
    asymp.LCL = asymp.LCL * -1,
    asymp.UCL = asymp.UCL * -1
  )


plot_data <- emm_data_2 |> mutate(dv = estimate)

g <- gg_rt_2_lmer(
  title = "Hemifield bias, squares only (EHI cut at +/-40)",
  plot_data,
  handedness_labeller = NULL,
  plot_colors = plot_colors,
  direction_labels = list(up = "LVF bias",
                          down = "RVF bias"),
  direction_labels_pos = list(up = 38,
                              down = -38),
  ylims = list(upper = 38,
               lower = -38),
  ybreaks = list(major = 10,
                 minor = 5),
  n_subjects = list(
    right = 378,
    left = 331
  )
)

ggsave(fig_path_var, g, "png", height = 3, width = 4)

include_graphics(fig_path_var)
```
<br>
```{r plot_cat_model_zoom_squares}
## Categorical plots: EHI extremes, squares only
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_lmer_zoom_squares.png")
fig_path_var_wide <- here(fig_dir, "rt_1_lmer_zoom_squares_wide.png")

rt_emm <- readRDS(here(model_cache_dir, "rt_emm_2bins.rds"))

## Get estimates for right and left handers.
zoom_squares_data <- aah_for_rt_model |>
  filter(handedness_extremes %in% c("Right", "Left")) |> 
  filter(shape == "Square")

rt_model <- lmer(data = zoom_squares_data,
                 rt ~ field * level * handedness + (1 |subject)
            )

rt_emm <- rt_model |> emmeans(~field * level * handedness)

emm_data_2 <- rt_emm |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95) |>
  as_tibble() |>
  ## Fix units
  mutate(
    estimate = estimate * -1,
    asymp.LCL = asymp.LCL * -1,
    asymp.UCL = asymp.UCL * -1
  )


plot_data <- emm_data_2 |> mutate(dv = estimate)

g <- gg_rt_2_lmer(
  title = "Hemifield bias, squares only (EHI +/-100)",
  plot_data,
  handedness_labeller = NULL,
  plot_colors = plot_colors,
  direction_labels = list(up = "LVF bias",
                          down = "RVF bias"),
  direction_labels_pos = list(up = 38,
                              down = -40),
  ylims = list(upper = 38,
               lower = -38),
  ybreaks = list(major = 10,
                 minor = 5),
    n_subjects = list(
    right = 182,
    left = 138
  )
)

ggsave(fig_path_var, g, "png", height = 3, width = 4)

include_graphics(fig_path_var)
```
<br>

## Preregistered 3-way (RT, continuous) {.tabset}

```{r plot_rt_cor_bins_config}
# plot_color = plot_blue
middle_purple1 <- "#996F91" ## In between plot colors one and two
middle_purple2 <- "#886e92" ## In between plot colors one and two
plot_color <- middle_purple1

h_plot_colors <- c(plot_colors[[1]], middle_purple1, plot_colors[[2]])
```


### Subject-level means and SEM


```{r plot_rt_1_cor_bins_line}
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_cor_bins_line.png")
fig_path_var_wide <- here(fig_dir, "rt_1_cor_bins_line.png")


## Extract line of best fit, confidence bounds from model.
model_cache_dir <- here("manual_cache", "exp_n1008_b")
rt_model_ehi <- readRDS(here(model_cache_dir, "rt_model_ehi.rds"))
ref <- ref_grid(rt_model_ehi, at = list(ehi = seq(-100, 100, 12.5)))
line_data <- ref |> contrast(interaction = c("consec"), by = "ehi", level = .95) |>
  summary(infer = T) |>
  as_tibble()

rt_subject_plot <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  left_join(aah_summary, by = c("subject", "handedness")) |>
  select(subject, ehi, starts_with("LVF"), handedness, handedness_extremes) |>
  ## Set "dv" here to make it easy to write a reusable function
  mutate(dv = LVF_Global_Bias) |>
  ungroup()


if (use_cached_rtfigs == F) {
  
  g <-
    gg_rt_1_line(
      title = "Frequency specialization by handedness (continuous)",
      y_title = "RVF - LVF, Local - Global reaction time (ms)",
      rt_subject_plot = rt_subject_plot,
      plot_colors = h_plot_colors,
      plot_color = plot_color,
      direction_labels = list(
        up = "LVF Global Bias",
        down = "RVF Global bias"),
      direction_labels_pos = list(
        up = 45,
        down = -15),
      ylims = list(
        upper = 46,
        lower = -30),
      ybreaks = list(
        major = 10,
        minor = 5)
    ) |>
    gg_style_cor_bin()
  
  g <- g +
  geom_line(data = line_data,
            aes(y = estimate)) +
    geom_ribbon(data = line_data,
                inherit.aes = F,
              aes(x = ehi, ymin = asymp.LCL, ymax = asymp.UCL),
              fill = plot_color,
              alpha = .2)
  
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  ggsave(fig_path_var_wide, g, "png", height = 4, width = 8)
  
}

include_graphics(fig_path_var)
```


```{r plot_cor_bins_line_shapes_only}
### Make a correlation plot for shapes only

## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_cor_bins_line_squares.png")
fig_path_var_wide <- here(fig_dir, "rt_1_cor_bins_line_squares.png")


## Extract line of best fit, confidence bounds from model.
squares_data <- aah_for_rt_model |> filter(shape == "Square")
  
if (use_cached_models == FALSE) {
  rt_model_ehi <- lme4::lmer(rt ~ field*level*ehi + (1|subject), data = squares_data)
  ## Create emmeans model object, and manually cache it.
  rt_emm_ehi <- emmeans(rt_model_ehi, ~ field * level * ehi)
  
  ## Manually cache model
  saveRDS(rt_model_ehi, here(manual_cache_dir, "rt_model_ehi_squares.rds"))
  saveRDS(rt_emm_ehi, here(manual_cache_dir, "rt_emm_ehi_squares.rds"))
  
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_model_ehi <- readRDS(here(manual_cache_dir, "rt_model_ehi_squares.rds"))
  rt_emm_ehi <- readRDS(here(manual_cache_dir, "rt_emm_ehi_squares.rds"))
}


ref <- ref_grid(rt_model_ehi, at = list(ehi = seq(-100, 100, 12.5)))
line_data <- ref |> contrast(interaction = c("consec"), by = "ehi", level = .95) |>
  summary(infer = T) |>
  as_tibble()

rt_subject_plot_squares <-  aah_correct |>
  filter(shape == "Square") |> 
  group_by(subject, field, level, handedness) |>
  summarize(rt = median(rt)) |> 
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  left_join(aah_summary, by = c("subject", "handedness")) |>
  select(subject, ehi, starts_with("LVF"), handedness, handedness_extremes) |>
  ## Set "dv" here to make it easy to write a reusable function
  mutate(dv = LVF_Global_Bias) |>
  ungroup()


if (use_cached_rtfigs == F) {
  
  g <-
    gg_rt_1_line(
      title = "Frequency specialization by handedness (continuous), squares only",
      y_title = "RVF - LVF, Local - Global reaction time (ms)",
      rt_subject_plot = rt_subject_plot_squares,
      plot_colors = h_plot_colors,
      plot_color = plot_color,
      direction_labels = list(
        up = "LVF Global Bias",
        down = "RVF Global bias"),
      direction_labels_pos = list(
        up = 65,
        down = -15),
      ylims = list(
        upper = 70,
        lower = -35),
      ybreaks = list(
        major = 10,
        minor = 5)
    ) |>
    gg_style_cor_bin()
  
  g <- g +
  geom_line(data = line_data,
            aes(y = estimate)) +
    geom_ribbon(data = line_data,
                inherit.aes = F,
              aes(x = ehi, ymin = asymp.LCL, ymax = asymp.UCL),
              fill = plot_color,
              alpha = .2)
  g
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  ggsave(fig_path_var_wide, g, "png", height = 4, width = 8)
  
}

include_graphics(fig_path_var)
```


### Rejects

Dots only (no lines)

```{r plot_rt_1_cor_bins}
fig_path_var <- here(fig_dir, "rt_1_cor_bins.png")
fig_path_var_wide <- here(fig_dir, "rt_1_cor_bins.png")

rt_subject_plot <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  left_join(aah_summary, by = c("subject", "handedness")) |>
  select(subject, ehi, starts_with("LVF"), handedness, handedness_extremes) |> 
  ## Set "dv" here to make it easy to write a reusable function
  mutate(dv = LVF_Global_Bias) |> 
  ungroup()

title = "Global bias asymmetry by EHI"
y_title = "RVF - LVF, Local - Global reaction time (ms)"
rt_subject_plot = rt_subject_plot
n_bins = 17
plot_colors = h_plot_colors
plot_color = plot_color
  
if (use_cached_rtfigs == F) {
  
  g <-
    gg_rt_1_ntiles(
      title = "Global bias asymmetry by handedness (continuous)",
      y_title = "RVF - LVF, Local - Global reaction time (ms)",
      rt_subject_plot = rt_subject_plot,
      n_bins = 17,
      plot_colors = h_plot_colors,
      plot_color = middle_purple1,
      direction_labels = list(
        up = "LVF Global Bias",
        down = "RVF Global bias"),
      direction_labels_pos = list(
        up = 45,
        down = -35),
      ylims = list(
        upper = 50,
        lower = -40),
      ybreaks = list(
        major = 20,
        minor = 10)
    ) |>
    gg_style_cor_bin() +
    ## Manually override tick mark labels (for 17-bin graph)
    scale_x_discrete(
      labels = c(
        "-100", "", "-75", "",
        "- 50", "", "-25", "",
        "0", "",
        "25", "", "50", "",
        "75", "", "100", "")
    )
  
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  ggsave(fig_path_var_wide, g, "png", height = 4, width = 8)
  
}

include_graphics(fig_path_var_wide)
```

```{r plot_rt_2_cor_bins}
fig_path_var <- here(fig_dir, "rt_2_cor_bins.png")


rt_subject_plot <- rt_subject |>
  pivot_wider(names_from = c(field),
              values_from = rt) |>
   mutate(LVF_Bias = RVF - LVF) |> 
  left_join(aah_summary, by = c("subject", "handedness")) |>
  select(subject, ehi, starts_with("LVF"), handedness, handedness_extremes) |> 
  ## Set "dv" here to make it easy to write a reusable function
  mutate(dv = LVF_Bias) |> 
  ungroup()

title = "Global bias asymmetry by EHI"
y_title = "RVF - LVF, Local - Global reaction time (ms)"
rt_subject_plot = rt_subject_plot
n_bins = 17
plot_colors = h_plot_colors
plot_color = plot_color
  
if (use_cached_rtfigs == F) {
  
  g <-
    gg_rt_1_ntiles(
      title = "Hemifield bias for global and local targets, by EHI",
      y_title = "Difference in reaction time between LVF and RVF",
      rt_subject_plot = rt_subject_plot,
      n_bins = 17,
      plot_colors = h_plot_colors,
      plot_color = middle_purple1,
      group_by_level = T, ## Hack to quickly make 2-box plot
      direction_labels = list(
        up = "LVF Global Bias",
        down = "RVF Global bias"),
      direction_labels_pos = list(
        up = 45,
        down = -35),
      ylims = list(
        upper = 50,
        lower = -40),
      ybreaks = list(
        major = 20,
        minor = 10)
    ) |>
    gg_style_cor_bin() +
    ## Manually override tick mark labels (for 17-bin graph)
    scale_x_discrete(
      labels = c(
        "-100", "", "-75", "",
        "- 50", "", "-25", "",
        "0", "",
        "25", "", "50", "",
        "75", "", "100", "")
    ) + 
    ## Add level to make 2-box plot
    facet_grid(cols = vars(level)) +
    scale_y_continuous(
      expand = expansion(mult = c(.025, .025))
    )
  
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
  
}

include_graphics(fig_path_var)
```

#### A dot for every subject

```{r plot_rt_1_cor_dots}
fig_path_var <- here(fig_dir, "rt_1_cor_dots.png")
fig_path_var_wide <- here(fig_dir, "rt_1_cor_dots_wide.png")

if (use_cached_rtfigs == F) {
  
  rt_subject_plot <- rt_subject |>
    pivot_wider(names_from = c(field, level),
                values_from = rt) |>
    mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
    left_join(aah_summary, by = c("subject", "handedness")) |>
    select(subject, ehi, starts_with("LVF"), handedness, handedness_extremes)
  
  g <- gg_rt_1_cor(title = "Global bias asymmetry by EHI",
              rt_subject_plot = rt_subject_plot,
              plot_colors = h_plot_colors)
  
  
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  ggsave(fig_path_var_wide, g, "png", height = 4, width = 8)
  
}

include_graphics(fig_path_var_wide)
```

```{r plot_rt_2_cor_dots}
fig_path_var <- here(fig_dir, "rt_2_cor_dots.png")

if (use_cached_rtfigs == F) {
  
  rt_subject_plot <- rt_subject |>
    pivot_wider(names_from = c(field),
                values_from = rt) |> 
    mutate(LVF_Bias = RVF - LVF) |> 
    left_join(aah_summary, by = c("subject", "handedness")) |>
    select(subject, ehi, level, starts_with("LVF"), handedness, handedness_extremes)
  
  g <- gg_rt_2_cor(title = "Hemifield bias for global and local targets, by EHI",
              rt_subject_plot = rt_subject_plot,
              plot_colors = h_plot_colors)
  
  ggsave(fig_path_var, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```


I think the dot for every subject is not very useful! I think the summary stats for LVF Global bias (top graph) help show the effect -- the group of left handers on the far left really pops out. Coloring by categorical group could make it even clearer. And, it could be nice to superimpose a trend line (from the subject level data pictured?) 

It also could be worth visualizing the summary stats from the model in this way, with the point estimates from the full mixed model.

<!-- ## Preregistered 3-way (accuracy, continuous) -->


## Discussion

### Questions for discussion

- Should we show SEM/CI from the subject-level data, or from the full mixed model?
- For the categorical plots, would it be more compelling and clear to visualize summary stats only, leaving out the subject-level dots? I like to "show all the data", but in this case, the wide range makes the patterns too hard to see.
- For the categorical plots, would it be more compelling and clear to visualize summary stats from the model, instead of/in addition to the subject-level data and subject level summary stats (which aren't the estimates we will discuss in Results)? 
- Is it enough to show the "tier one" figures listed below? Should we show any figures illustrating zoom analyses, and the continuous analysis?
- Which figures show the interaction most clearly? Could we just show the 4-box plot?
- Is it worth showing the 8-box data for the main analysis? I think it would be nice to show this more "raw" data, to help people understand the design. 
- Any feedback on the design of the graphs?
- Should we show the accuracy (null) results?


### Proposal

#### Tier one
(1) EHI
(2) (a, b, c) Main result: Compound figure showing RT interaction (1-box, 4-box, 8-box) in full sample.

#### Tier two (could leave out if we need the space)
(3) (a, b) Zoom: compound figure showing effect of shape, within strong lefties (1-box, 4-box). This shows the stro
(4) Scatterplot-like figure showing RT interaction (continuous) [If convincing]

### Brainstorm - all figures I think might be worth showing
- EHI
- (a, b, c) Main result: Compound figure showing RT interaction (1-box, 4-box, 8-box) in full sample.
- (a, b) Zoom: Compound figure showing RT interaction (1-box, 4-box) in strong handers, or strong handers with squares only.
- (a, b) Zoom: show effect of shape (1-box, 4-box),  either in strong handers only (revealing largest effect size, and also illustrating the effect of shape), or in the full sample (focusing on the effect of shape itself)
- Scatterplot-like figure showing RT interaction (continuous)
